==> 04_manuscript_figures.Rmd <==
---
title: "IOH and dengue manuscript figures"
author: "Alyson Singleton"
date: "`r Sys.Date()`"
output:
  html_document: default
  word_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
# install packages, load libraries, set themes, knit formatting
pacman::p_load(tidyverse, dplyr, kableExtra, knitr, sf, raster, terra, spData, tmap, leaflet, ggplot2, spThin, mapdata, rnaturalearth, rnaturalearthdata, brazilmaps, devtools, brazilmaps, mapview, grid, gridExtra, RColorBrewer, raster, plm, fixest, cobalt, broom, ggpubr, gridExtra, cowplot, ggspatial, scales)
theme_set(theme_bw())
knitr::opts_chunk$set(echo = F, results='asis', warning=F, message=F, cache=F,
                      fig.height=5, fig.width=10)

get_legend<-function(myggplot){
  tmp <- ggplot_gtable(ggplot_build(myggplot))
  leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
  legend <- tmp$grobs[[leg]]
  return(legend)
}

library(showtext)
font_add("Arial", "/Library/Fonts/Arial.ttf")  # Use the actual file path
showtext_auto()

```

```{r, load yearly case data}
vert_line_date <- as.Date('2008-01-01')

###################
### dengue data ###
###################
### yearly data
dengue_df_yearly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_yearly_full_dataset_c.csv")
dengue_df_yearly$year <- as.factor(dengue_df_yearly$year)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_yearly$incidence <- (dengue_df_yearly$yearly_cases+1)/dengue_df_yearly$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_yearly$urban <- dengue_df_yearly$urban + 0.001
dengue_df_yearly$ag <- dengue_df_yearly$ag + 0.001

# double check complete cases
dengue_df_yearly <- dengue_df_yearly[complete.cases(dengue_df_yearly),]

# remove extreme outliers (all most likely due to inaccurate pop data)
dengue_df_yearly <- dengue_df_yearly[which(dengue_df_yearly$incidence < 0.5),]

# check how many are always zero
dengue_df_yearly_zeroes <- dengue_df_yearly %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(yearly_cases))
clusters_w_at_least_one_case <- dengue_df_yearly$cluster[which(dengue_df_yearly_zeroes$total_cases>0)]
dengue_df_yearly_at_least_one_case <- dengue_df_yearly[which(dengue_df_yearly$cluster %in% clusters_w_at_least_one_case),]

# remove PM
dengue_df_yearly_no_pm <- dengue_df_yearly[!(dengue_df_yearly$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_yearly_buffered <- dengue_df_yearly[which(dengue_df_yearly$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
dengue_df_yearly_buffered_no_pm <- dengue_df_yearly_buffered[!(dengue_df_yearly_buffered$cluster %in% 1),]

###################
### leish data ####
###################
### yearly data
leish_df_yearly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/leish_yearly_full_dataset_c.csv")
leish_df_yearly$year <- as.factor(leish_df_yearly$year)

# cases + 1 so 0's dont go to infinity when logged
leish_df_yearly$incidence <- (leish_df_yearly$yearly_cases+1)/leish_df_yearly$population

# land use vars + 0.001 so 0's dont go to infinity when logged
leish_df_yearly$urban <- leish_df_yearly$urban + 0.001
leish_df_yearly$ag <- leish_df_yearly$ag + 0.001

# double check complete cases
leish_df_yearly <- leish_df_yearly[complete.cases(leish_df_yearly),]

# remove extreme outliers
leish_df_yearly <- leish_df_yearly[which(leish_df_yearly$incidence < 0.5),]

# check how many are always zero
leish_df_yearly_zeroes <- leish_df_yearly  %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(yearly_cases))
leish_clusters_w_at_least_one_case <- leish_df_yearly$cluster[which(leish_df_yearly_zeroes$total_cases>0)]
leish_df_yearly_at_least_one_case <- leish_df_yearly[which(leish_df_yearly$cluster %in% leish_clusters_w_at_least_one_case),]

# remove PM
leish_df_yearly_no_pm <- leish_df_yearly[!(leish_df_yearly$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
leish_df_yearly_buffered <- leish_df_yearly[which(leish_df_yearly$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
leish_df_yearly_buffered_no_pm <- leish_df_yearly_buffered[!(leish_df_yearly_buffered$cluster %in% 1),]
```

```{r, load biannual case data}
vert_line_date <- as.Date('2008-01-01')

###################
### dengue data ###
###################
### biannual data
dengue_df_biannual <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_biannual_full_dataset_c.csv")
dengue_df_biannual$biannual_date <- as.factor(dengue_df_biannual$biannual_date)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_biannual$incidence <- (dengue_df_biannual$biannual_cases+1)/dengue_df_biannual$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_biannual$urban <- dengue_df_biannual$urban + 0.001
dengue_df_biannual$ag <- dengue_df_biannual$ag + 0.001

# double check complete cases
dengue_df_biannual <- dengue_df_biannual[complete.cases(dengue_df_biannual),]

# remove extreme outliers
dengue_df_biannual <- dengue_df_biannual[which(dengue_df_biannual$incidence < 0.5),]

#check how many are always zero
dengue_df_biannual_zeroes <- dengue_df_biannual %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(biannual_cases))
clusters_w_at_least_one_case_biannual <- dengue_df_biannual$cluster[which(dengue_df_biannual_zeroes$total_cases>0)]
dengue_df_biannaul_at_least_one_case <- dengue_df_biannual[which(dengue_df_biannual$cluster %in% clusters_w_at_least_one_case_biannual),]

#remove PM
dengue_df_biannual_no_pm <- dengue_df_biannual[!(dengue_df_biannual$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_biannual_buffered <- dengue_df_biannual[which(dengue_df_biannual$all_cutoffs %in% c(1,2,4,5,6,7,0)),]

###################
### leish data ####
###################

### biannaul data
leish_df_biannual <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/leish_biannual_full_dataset_c.csv")
leish_df_biannual$biannual_date <- as.factor(leish_df_biannual$biannual_date)

# cases + 1 so 0's dont go to infinity when logged
leish_df_biannual$incidence <- (leish_df_biannual$biannual_cases+1)/leish_df_biannual$population

# land use vars + 0.001 so 0's dont go to infinity when logged
leish_df_biannual$urban <- leish_df_biannual$urban + 0.001
leish_df_biannual$ag <- leish_df_biannual$ag + 0.001

# double check complete cases
leish_df_biannual <- leish_df_biannual[complete.cases(leish_df_biannual),]

# remove extreme outliers
leish_df_biannual <- leish_df_biannual[which(leish_df_biannual$incidence < 0.5),]

# check how many are always zero
leish_df_biannual_zeroes <- leish_df_biannual  %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(biannual_cases))
leish_clusters_w_at_least_one_case_biannual <- leish_df_biannual$cluster[which(leish_df_biannual_zeroes$total_cases>0)]
leish_df_biannual_at_least_one_case <- leish_df_biannual[which(leish_df_biannual$cluster %in% leish_clusters_w_at_least_one_case_biannual),]

# remove PM
leish_df_biannual_no_pm <- leish_df_biannual[!(leish_df_biannual$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
leish_df_biannual_buffered <- leish_df_biannual[which(leish_df_biannual$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
```

```{r, Fig1}
#####################
## Fig 1a
#####################

cluster_ids <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_7.5km/cluster_centroids_7.5km.csv")
colnames(cluster_ids)[1] <- "cluster"
center_lat_long <- left_join(dengue_df_yearly[,c(3,20)], cluster_ids, by='cluster')

center_lat_long <- center_lat_long %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))

peru_depts <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/per_admbnda_adm1_ign_20200714.shp")
peru_depts <- st_as_sf(peru_depts) 
peru_depts$geometry <- st_transform(peru_depts$geometry, 4326)
peru_outline <- st_union(peru_depts$geometry)

districts <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/Madre_de_Dios.shp")
districts <- st_as_sf(districts) 
districts$geometry <- st_transform(districts$geometry, 4326)
mdd_region <- st_union(districts$geometry)

rivers <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_rivers2.shp")
rivers <- st_as_sf(rivers, crs=4326) 
rivers$geometry <- st_transform(rivers$geometry, "EPSG:4326")

roads <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_roads.shp")
roads <- st_as_sf(roads) 
roads$geometry <- st_transform(roads$geometry, 4326)
roads_mdd <- st_covers(mdd_region,roads$geometry, sparse = FALSE)
roads_mdd <- roads[roads_mdd[1,],]

highway <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/peru_roads_important.shp")
highway <- highway[which(highway$ref=="PE-30C"),]
highway <- st_as_sf(highway) 
highway$geometry <- st_transform(highway$geometry, 4326)

highway_mdd <- st_covers(mdd_region,highway$geometry, sparse = FALSE)
highway_mdd <- highway[highway_mdd[1,],]

no_axis <- theme(axis.title=element_blank(),
                 axis.text=element_blank(),
                 axis.ticks=element_blank(),
                 panel.grid.major = element_blank())

# Build columns for plotting
center_lat_long_fig1 <- center_lat_long_fig %>%
  mutate(all_cutoffs = case_when(
    all_cutoffs %in% c('0', '4', '5', '6', '7') ~ '0',
    all_cutoffs %in% c('1', '2') ~ '1',
    TRUE ~ as.character(all_cutoffs)),
    all_cutoffs = factor(all_cutoffs, levels = c('1', '0', '3')),
    pm_indicator = if_else(cluster == 1, 24, 21),
    pm_indicator2 = if_else(cluster == 1, 3.2, 2.5)) %>%
  arrange(if_else(cluster == 1, 1, 0))

# Coordinates for labeling
cluster1_coords <- st_coordinates(center_lat_long_fig1 %>% filter(cluster == 1))
arrow_x <- cluster1_coords[1, "X"] + 0.08
arrow_y <- cluster1_coords[1, "Y"]
label_x <- arrow_x + 0.14
label_y <- arrow_y - 0.07

# Plot MdD
mdd_map <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#3b3b3b', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='Rivers'), linewidth=0.2, show.legend = "line") +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='Unpaved Roads'), linewidth=0.5, show.legend = "line") +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='Highway'), linewidth=0.7, show.legend = "line") +
  geom_sf(data = center_lat_long_fig1, 
          aes(geometry = geometry, 
              fill=all_cutoffs, 
              shape=factor(pm_indicator)), 
          size=center_lat_long_fig1$pm_indicator2, 
          alpha=1) +
  scale_fill_manual(name= "", values=c("1" = "#E04490", "0" = "#648FFF", "3" = "#ffffff"),
                    labels=c("Exposed (<5km)", "Unexposed (>10km)", "Buffer (removed)")) +
  scale_color_manual(name = "", values = c("Unpaved Roads" = '#a6a6a6', "Highway" = 'black',"Rivers" = 'lightblue')) +
  scale_shape_manual(values = c(`21` = 21, `24` = 24), guide = "none") +
  theme_minimal() +
  no_axis +
  guides(fill=guide_legend(override.aes=list(shape=21, color = "black", linetype = 0, size=2.5))) +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='bottom',
        legend.spacing.x = unit(0.05, 'cm'),
        legend.box.spacing = unit(0.05, 'cm'),
        legend.key.width = unit(0.4, "cm")) +
  annotation_scale(location = "bl",height = unit(0.1, "cm")) +
  annotation_north_arrow(location = "br",style = north_arrow_orienteering(text_size = 6), 
                         height = unit(0.7, "cm"), width = unit(0.7, "cm")) +
  annotate("segment", 
           x = label_x-0.02, y = label_y, 
           xend = arrow_x, yend = arrow_y,
           arrow = arrow(length = unit(0.17, "cm")), 
           color = "black") +
  annotate("text", 
           x = label_x, y = label_y, 
           label = "PM", 
           size = 3.4, fontface = "bold", hjust = 0, color = "black")
mdd_map
fig1_legend <- get_legend(mdd_map)
mdd_map <- mdd_map + theme(legend.position = "none")

fig1a <- ggdraw() + 
    draw_plot(ggplot() +
              geom_sf(data = peru_outline, fill=NA, color='#3b3b3b', size=.001, show.legend = FALSE) +
              geom_sf(data = peru_depts, fill=NA, color='#3b3b3b', size=.001, show.legend = FALSE) +
              geom_sf(data = mdd_region, fill='#cfcfcf', color='#3b3b3b', size=.001, show.legend = FALSE) +
              theme_minimal() +
              no_axis +
              theme(legend.text=element_text(size=12),
                    legend.title=element_text(size=14),
                    legend.position = "none"),
            0.65, 0.65, 0.3, 0.3) + 
  draw_plot(mdd_map,
            0, 0, 1, 1)
fig1a

#####################
## Fig 1b
#####################

dengue_raw_plotting <- dengue_df_yearly_buffered %>%
  group_by(fivekm, year) %>%
  summarize(new_incidence = sum(yearly_cases)/sum(population))

dengue_raw_plotting$year <- as.Date(dengue_raw_plotting$year)
dengue_raw_plotting$fivekm <- as.character(dengue_raw_plotting$fivekm)

fig1b <- ggplot(dengue_raw_plotting) +
  geom_line(aes(x=year, y=new_incidence, colour=fivekm)) +
  geom_vline(xintercept=vert_line_date,linetype='dashed') +
  ggtitle("Dengue incidence") +
  xlab("Year") + ylab("") + 
  scale_color_manual(name = "Exposure", values=c("#648FFF","#E04490"), labels=c('Far (>10km)', 'Near (<5km)'),) +
  theme_bw()+
  theme(plot.title = element_text(size=12, face="bold"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=14),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=10),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "none",
        strip.text.x = element_text(size = 10))
fig1b

#####################
## Fig 1c
#####################

leish_raw_plotting <- leish_df_yearly_buffered %>%
  group_by(fivekm, year) %>%
  summarize(new_incidence = sum(yearly_cases)/sum(population))

leish_raw_plotting$year <- as.Date(leish_raw_plotting$year)
leish_raw_plotting$fivekm <- as.character(leish_raw_plotting$fivekm)
fig1c <- ggplot(leish_raw_plotting) +
  geom_line(aes(x=year, y=new_incidence, colour=fivekm)) +
  geom_vline(xintercept=vert_line_date,linetype='dashed') +
  scale_y_continuous(labels = function(x) round(x, 3)) +
  ggtitle("Leish incidence") +
  xlab("Year") + ylab("") + 
  scale_color_manual(name = "Exposure", values=c("#648FFF","#E04490"), labels=c('Far (>10km)', 'Near (<5km)'),) +
  theme_bw()+
  theme(plot.title = element_text(size=12, face="bold"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=14),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=8),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=10),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "none",
        strip.text.x = element_text(size = 10))
fig1c

#####################
## Fig 1all
#####################

fig1all <- grid.arrange(fig1a, fig1b, fig1c, fig1_legend,
                        ncol = 4, nrow = 3,
                        layout_matrix = rbind(c(1,1,2,2),c(1,1,3,3),c(NA,4,4,NA)), 
                        heights=c(3,3,1))

fig1all <- as_ggplot(fig1all) +                                
  draw_plot_label(label = c("A", "B", "C"), size = 14,
                  x = c(0.01, 0.48, 0.48), y = c(1, 1, 0.57)) 
fig1all

ggsave("Fig1.pdf", plot=fig1all, path="~/Desktop/doctorate/ch2 mdd highway/manuscript_figures/", width = 10, height = 6, units="in", device = "pdf")

```

```{r, Fig2}
#####################
## Fig 2a
#####################
### yearly model
dengue_yearly_model <- fenegbin(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffered)
df <- as.data.frame(dengue_yearly_model$coeftable)
df <- df[1:22,]
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$year <- c(seq(as.Date("2000-01-01"), as.Date("2007-01-01"), by="year"),
             seq(as.Date("2009-01-01"), as.Date("2022-01-01"), by="year"))
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1

fig2a <- ggplot(df) +
  geom_hline(aes(yintercept=0), colour='red', linewidth=.4) +
  geom_errorbar(aes(x=year, ymax=upper, ymin=lower), width=0, linewidth=0.5) +
  geom_vline(aes(xintercept=as.Date("2008-01-01")), linetype='dashed', linewidth=0.4) +
  geom_point(aes(x=as.Date("2008-01-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(year, estimate), size=3, shape=21, fill='white') +
  xlab("Year") + ylab("% change\ndengue\nincidence\nby year\nrelative\nto 2008") + 
  theme_bw() +
  scale_y_continuous(
    trans = scales::trans_new(
      name = "arcsinh",
      transform = function(x) asinh(x),
      inverse = function(x) sinh(x)),
    labels = scales::percent_format(accuracy = 1),
    limits=c(-1.2,15)) +
  #scale_y_continuous(labels = scales::percent, limits=c(-1.2,15)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=12),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "right",
        strip.text.x = element_text(size = 12))
fig2a

#####################
## Fig 2b
#####################

### aggregated data (treatment)
dengue_df_agg_biannual <- dengue_df_biannual_buffered
dengue_df_agg_biannual <- dengue_df_agg_biannual[which(as.Date(dengue_df_biannual_buffered$biannual_date) > '2006-10-01'),]
dengue_df_agg_biannual$biannual_binary <- ifelse(as.Date(dengue_df_agg_biannual$biannual_date) > '2008-04-01',1, 0)
dengue_df_agg_biannual$month <- format(as.Date(dengue_df_agg_biannual$biannual_date), "%m")
dengue_df_agg_biannual_dry <- dengue_df_agg_biannual[which(dengue_df_agg_biannual$month == "04"),]
dengue_df_agg_biannual_rainy <- dengue_df_agg_biannual[which(dengue_df_agg_biannual$month == "10"),]

### aggregated models
dengue_biannual_agg_model_dry <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + biannual_date, vcov = "cluster", data = dengue_df_agg_biannual_dry)
summary(dengue_biannual_agg_model_dry)

dengue_biannual_agg_model_rainy <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + biannual_date, vcov = "cluster", data = dengue_df_agg_biannual_rainy)
summary(dengue_biannual_agg_model_rainy)

df <- as.data.frame(rbind(dengue_biannual_agg_model_dry$coeftable[5,],
                          dengue_biannual_agg_model_rainy$coeftable[5,]))
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
df$rainy <- c("Dry","Rainy")

fig2b <- ggplot(df) +
  geom_hline(aes(yintercept=0), colour='red', size=.4) +
  geom_errorbar(aes(x=rainy, ymax=upper, ymin=lower, colour=rainy), width=0, size=0.5) +
  geom_point(aes(rainy, estimate, fill=rainy), size=3, shape=21) +
  scale_fill_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  scale_colour_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  xlab("Season") + ylab("") + 
  theme_bw() +
  scale_y_continuous(
    trans = scales::trans_new(
      name = "arcsinh",
      transform = function(x) asinh(x),
      inverse = function(x) sinh(x)),
    labels = scales::percent_format(accuracy = 1),
    limits=c(-1.2,15)) +
  #scale_y_continuous(labels = scales::percent, limits=c(-1.2,15)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=12),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_blank(),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "",
        strip.text.x = element_text(size = 12))
fig2b

#####################
## Fig 2all
#####################

fig2all <- grid.arrange(fig2a, fig2b,                     
                        ncol = 2, nrow = 1,
                        layout_matrix = rbind(c(1,1,1,1,1,2)))#,heights=c(7,1))

fig2all <- as_ggplot(fig2all) +                                
  draw_plot_label(label = c("A", "B"), size = 17,
                  x = c(0.138, 0.846), y = c(0.99, 0.99)) 

fig2all

ggsave("Fig2.pdf", plot=fig2all, path="~/Desktop/doctorate/ch2 mdd highway/manuscript_figures/", width = 10, height = 6, units="in", device = "pdf")
```

```{r, attributable cases}
### aggregated data (treatment)
dengue_df_agg <- dengue_df_yearly_buffered
dengue_df_agg <- dengue_df_agg[which(as.Date(dengue_df_agg$year) > '2007-01-01'),] ##NOTE change this!**
dengue_df_agg$year_binary <- ifelse(as.Date(dengue_df_agg$year) > '2008-01-01',1,0)

### aggregated model
dengue_yearly_agg_model <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg)
summary(dengue_yearly_agg_model)

df <- as.data.frame(dengue_yearly_agg_model$coeftable)
df <- df[5,]
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$estimate <- exp(df$estimate)
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error

## build df w information needed for IRR/attributable cases calculations
incidence_df <- dengue_df_agg %>%
  group_by(year_binary,fivekm) %>%
  summarise(incidence_avg = mean(incidence),
            yearly_cases = sum(yearly_cases),
            pop_sum = sum(population)) %>%
  pivot_wider(names_from = fivekm, values_from = c(incidence_avg,yearly_cases,pop_sum))

# get IRR from model
model_IRR <- df$estimate #mean
model_IRR <- df$upper #max
model_IRR <- df$lower #min

# calculate attributable proportion, attributable number of cases, prop of total cases since paving
attr_prop <- ((model_IRR-1)/model_IRR)*
  (incidence_df$pop_sum_1[2]/(incidence_df$pop_sum_0[2]+incidence_df$pop_sum_1[2]))
attr_prop
attr_cases <- incidence_df$yearly_cases_1[2]*attr_prop
attr_cases
attr_cases/sum(incidence_df$yearly_cases_0[2]+incidence_df$yearly_cases_1[2])

# percentage change
df <- as.data.frame(dengue_yearly_agg_model$coeftable)
df <- df[5,]
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
df
```

```{r, Fig3}
#####################
## Fig 3a
#####################
### yearly leish model
leish_yearly_model <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + sum_precip + mean_temp + log(urban) + log(ag) | cluster + year, vcov = "cluster", data = leish_df_yearly_buffered)
#summary(leish_yearly_model)

df <- as.data.frame(leish_yearly_model$coeftable)
df <- df[1:22,]
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$year <- c(seq(as.Date("2000-01-01"), as.Date("2007-01-01"), by="year"),
             seq(as.Date("2009-01-01"), as.Date("2022-01-01"), by="year"))
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1

fig3a <- ggplot(df) +
  geom_hline(aes(yintercept=0), colour='red', size=.4) +
  geom_errorbar(aes(x=year, ymax=upper, ymin=lower), width=0, size=0.5) +
  geom_vline(aes(xintercept=as.Date("2008-01-01")), linetype='dashed', size=0.4) +
  geom_point(aes(x=as.Date("2008-01-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(year, estimate), size=3, shape=21, fill='white') +
  xlab("Year") + ylab("% change\nleish\nincidence\nby year\nrelative\nto 2008") + 
  theme_bw()+
  scale_y_continuous(
    trans = scales::trans_new(
      name = "arcsinh",
      transform = function(x) asinh(x),
      inverse = function(x) sinh(x)),
    labels = scales::percent_format(accuracy = 1),
    limits=c(-1.2,15)) +
  # scale_y_continuous(labels = scales::percent, limits=c(-1.2,15)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=18),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "right",
        strip.text.x = element_text(size = 12))
fig3a

#####################
## Fig 3b
#####################

### aggregated data (treatment)
leish_df_agg_biannual <- leish_df_biannual_buffered
leish_df_agg_biannual <- leish_df_agg_biannual[which(as.Date(leish_df_biannual_buffered$biannual_date) > '2006-10-01'),]
leish_df_agg_biannual$biannual_binary <- ifelse(as.Date(leish_df_agg_biannual$biannual_date) > '2008-04-01',1, 0)
leish_df_agg_biannual$month <- format(as.Date(leish_df_agg_biannual$biannual_date), "%m")
leish_df_agg_biannual_dry <- leish_df_agg_biannual[which(leish_df_agg_biannual$month == "04"),]
leish_df_agg_biannual_rainy <- leish_df_agg_biannual[which(leish_df_agg_biannual$month == "10"),]

### aggregated models
leish_biannual_agg_model_dry <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + biannual_date, vcov = "cluster", data = leish_df_agg_biannual_dry)
summary(leish_biannual_agg_model_dry)

leish_biannual_agg_model_rainy <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + biannual_date, vcov = "cluster", data = leish_df_agg_biannual_rainy)
summary(leish_biannual_agg_model_rainy)

df <- as.data.frame(rbind(leish_biannual_agg_model_dry$coeftable[5,],
                          leish_biannual_agg_model_rainy$coeftable[5,]))
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
df$rainy <- c("Dry","Rainy")

fig3b <- ggplot(df) +
  geom_hline(aes(yintercept=0), colour='red', size=.4) +
  geom_errorbar(aes(x=rainy, ymax=upper, ymin=lower, colour=rainy), width=0, size=0.5) +
  geom_point(aes(rainy, estimate, fill=rainy), size=3, shape=21) +
  scale_fill_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  scale_colour_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  xlab("Season") + ylab("") + 
  theme_bw()+
  scale_y_continuous(
    trans = scales::trans_new(
      name = "arcsinh",
      transform = function(x) asinh(x),
      inverse = function(x) sinh(x)),
    labels = scales::percent_format(accuracy = 1),
    limits=c(-1.2,15)) +
  # scale_y_continuous(labels = scales::percent, limits=c(-1.2,15)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=12),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_blank(),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "",
        strip.text.x = element_text(size = 12))
fig3b

#####################
## Fig 3all
#####################
fig3all <- grid.arrange(fig3a, fig3b,                     
                        ncol = 2, nrow = 1,
                        layout_matrix = rbind(c(1,1,1,1,1,2)))#,heights=c(7,1))

fig3all <- as_ggplot(fig3all) +                                
  draw_plot_label(label = c("A", "B"), size = 17,
                  x = c(0.138, 0.846), y = c(0.99, 0.99)) 

fig3all

ggsave("Fig3.pdf", plot=fig3all, path="~/Desktop/doctorate/ch2 mdd highway/manuscript_figures/", width = 10, height = 6, units="in", device = "pdf")
```

```{r, Fig4}
#####################
## Fig 4a
#####################

### aggregated data (treatment)
dengue_df_agg <- dengue_df_yearly
dengue_df_agg <- dengue_df_agg[which(as.Date(dengue_df_agg$year) > '2007-01-01'),]
dengue_df_agg$year_binary <- ifelse(as.Date(dengue_df_agg$year) > '2008-01-01',1,0)

## bootstrap to get more conservative standard errors
treatment_col_names <- c('onekm','fivekm','tenkm','fifteenkm','twentykm','thirtykm')
std_error_stor <- c()
for(j in 1:5){
  #build control and treatment groups for each distance subgroup
  dengue_df_yearly_0s <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(6,7,0)),] #same control group each time
  dengue_df_yearly_1s <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(j)),] #loop through each distance group
  
  for(k in 1:1000){
    # sample cluster ids
    set.seed(k)
    dengue_df_yearly_sample0s <- sample(dengue_df_yearly_0s$cluster,length(unique(dengue_df_yearly_0s$cluster)),replace=T)
    dengue_df_yearly_sample1s <- sample(dengue_df_yearly_1s$cluster,length(unique(dengue_df_yearly_1s$cluster)),replace=T)
    ids_to_pull <- c(dengue_df_yearly_sample0s,dengue_df_yearly_sample1s)
    
    # build new dataset
    dengue_df_yearly_boot <- c()
    for(i in ids_to_pull) {
      og_rows_associated_w_id_i <- dengue_df_agg[which(dengue_df_agg$cluster == i),]
      dengue_df_yearly_boot <- rbind(dengue_df_yearly_boot, og_rows_associated_w_id_i)
    }
    
    #run model
    col_name <- treatment_col_names[j]
    dengue_yearly_model_boot <- fepois(as.formula(paste0("incidence ~ year_binary*", col_name, "+ log(urban) + log(ag) + sum_precip + mean_temp | cluster + year")), 
                                           vcov = "cluster", 
                                           data = dengue_df_yearly_boot)
    
    #store standard error
    std_error <- coeftable(dengue_yearly_model_boot)[5,2]
    std_error_stor <- c(std_error_stor,std_error)
  }
}

# create datasets to run aggregated models
dengue_df_yearly_1km <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(1,6,7,0)),]
dengue_df_yearly_5km <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(2,6,7,0)),]
dengue_df_yearly_10km <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(3,6,7,0)),]
dengue_df_yearly_15km <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(4,6,7,0)),]
dengue_df_yearly_20km <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(5,6,7,0)),]

# run models
dengue_yearly_model_1km <- fepois(incidence ~ year_binary*onekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_1km)
dengue_yearly_model_5km <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_5km)
dengue_yearly_model_10km <- fepois(incidence ~ year_binary*tenkm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_10km)
dengue_yearly_model_15km <- fepois(incidence ~ year_binary*fifteenkm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_15km)
dengue_yearly_model_20km <- fepois(incidence ~ year_binary*twentykm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_20km)

df <- as.data.frame(rbind(dengue_yearly_model_1km$coeftable[5,],
                          dengue_yearly_model_5km$coeftable[5,],
                          dengue_yearly_model_10km$coeftable[5,],
                          dengue_yearly_model_15km$coeftable[5,],
                          dengue_yearly_model_20km$coeftable[5,]))

df$Cutoff <- c('1km','5km','10km', '15km','20km')
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value', 'Cutoff')
df$Cutoff <- factor(as.character(df$Cutoff), levels=c('1km','5km','10km', '15km','20km'))
df$estimate <- as.numeric(df$estimate)

# replace analytic standard errors w bootstrapped estimates
df$std_error_boot <- c(mean(std_error_stor[1:1000]),
                       mean(std_error_stor[1001:2000]),
                       mean(std_error_stor[2001:3000]),
                       mean(std_error_stor[3001:4000]),
                       mean(std_error_stor[4001:5000]))

df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1

fig4a <- ggplot(df, aes(group=Cutoff)) +
  geom_hline(aes(yintercept=0), colour='red', linewidth=.4) +
  geom_errorbar(aes(x=Cutoff, ymax=upper, ymin=lower, color=Cutoff), width=0, linewidth=0.7) +
  geom_point(aes(Cutoff, estimate, fill=Cutoff), color='black', size=3, shape=21) +
  scale_color_manual(name="Treatment", values=list('1km'="#DC267F",'5km'='#FF7900','10km'='#FFB000',
                                                   '15km'='#648FFF','20km'='#785EF0','30km'='#004600'))+ 
  scale_fill_manual(name="Treatment", values=list('1km'="#DC267F",'5km'='#FF7900','10km'='#FFB000',
                                                  '15km'='#648FFF','20km'='#785EF0','30km'='#004600'))+ 
  xlab("") + ylab("% change\ndengue\nincidence\nrelative\nto 2008") + 
  theme_bw()+
  scale_y_continuous(
    trans = scales::trans_new(
      name = "arcsinh",
      transform = function(x) asinh(x),
      inverse = function(x) sinh(x)),
    labels = scales::percent_format(accuracy = 1)) +
  #scale_y_continuous(labels = scales::percent, limits=c(-1,18)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=14),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 12))

fig4a

#####################
## Fig 4b
#####################

boundary_dummy_vars <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/boundary_dummy_vars_7.5km/boundary_dummy_vars_7.5km.csv")
cluster_ids <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_7.5km/cluster_centroids_7.5km.csv")
colnames(cluster_ids)[1] <- "cluster"
center_lat_long <- left_join(dengue_df_yearly, cluster_ids, by='cluster')

center_lat_long <- center_lat_long %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))

districts <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/Madre_de_Dios.shp")
districts <- st_as_sf(districts) 
districts$geometry <- st_transform(districts$geometry, 4326)
mdd_region <- st_union(districts$geometry)

rivers <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_rivers2.shp")
rivers <- st_as_sf(rivers, crs=4326) 
rivers$geometry <- st_transform(rivers$geometry, "EPSG:4326")

roads <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_roads.shp")
roads <- st_as_sf(roads) 
roads$geometry <- st_transform(roads$geometry, 4326)
roads_mdd <- st_covers(mdd_region,roads$geometry, sparse = FALSE)
roads_mdd <- roads[roads_mdd[1,],]

highway <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/peru_roads_important.shp")
highway <- highway[which(highway$ref=="PE-30C"),]
highway <- st_as_sf(highway) 
highway$geometry <- st_transform(highway$geometry, 4326)

highway_mdd <- st_covers(mdd_region,highway$geometry, sparse = FALSE)
highway_mdd <- highway[highway_mdd[1,],]

no_axis <- theme(axis.title=element_blank(),
                 axis.text=element_blank(),
                 axis.ticks=element_blank(),
                 panel.grid.major = element_blank())

# Build columns for plotting
center_lat_long_fig4 <- center_lat_long %>%
  mutate(all_cutoffs = case_when(
    all_cutoffs %in% c('0', '6', '7') ~ '0',
    TRUE ~ as.character(all_cutoffs)),
    all_cutoffs = factor(all_cutoffs, levels = c('1','2','3','4','5','0')),
    pm_indicator = if_else(cluster == 1, 24, 21),
    pm_indicator2 = if_else(cluster == 1, 3.2, 2.5)) %>%
  arrange(if_else(cluster == 1, 1, 0))

# Coordinates for labeling
cluster1_coords <- st_coordinates(center_lat_long_fig1 %>% filter(cluster == 1))
arrow_x <- cluster1_coords[1, "X"] + 0.08
arrow_y <- cluster1_coords[1, "Y"]
label_x <- arrow_x + 0.14
label_y <- arrow_y - 0.07

# Plot MdD
fig4b <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#a6a6a6', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='lightblue'), linewidth=0.2, show.legend = FALSE) +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='#a6a6a6'), linewidth=0.5, show.legend = FALSE) +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='black'), linewidth=0.7, show.legend = FALSE) +
  geom_sf(data = center_lat_long_fig4, 
          aes(geometry = geometry, 
              fill=all_cutoffs, 
              shape=factor(pm_indicator)), 
          size=center_lat_long_fig1$pm_indicator2, 
          stroke=0.5,
          alpha = 1) +
  scale_fill_manual(name= "", values=list('1'="#DC267F",'2'='#FF7900','3'='#FFB000',
                                          '4'='#648FFF','5'='#785EF0','0'='lightgrey'), 
                    labels=c("1km", "5km", "10km", "15km", "20km", "Control")) +
  scale_color_manual(name="", values=c('#a6a6a6','black','lightblue'),
                     labels=c('Unpaved Roads','Highway','Rivers')) +
  scale_shape_manual(values = c(`21` = 21, `24` = 24), guide = "none") +
  theme_minimal() +
  no_axis +
  guides(fill=guide_legend(override.aes=list(shape=21, color = "black", linetype = 0, size=2.5))) +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='bottom',
        legend.spacing.x = unit(0.05, 'cm'),
        legend.box.spacing = unit(0.05, 'cm'),
        legend.key.width = unit(0.4, "cm")) +
  annotation_scale(location = "bl",height = unit(0.1, "cm")) +
  annotation_north_arrow(location = "br",style = north_arrow_orienteering(text_size = 6), 
                         height = unit(0.7, "cm"), width = unit(0.7, "cm")) +
  annotate("segment", 
           x = label_x-0.02, y = label_y, 
           xend = arrow_x, yend = arrow_y,
           arrow = arrow(length = unit(0.17, "cm")), 
           color = "black") +
  annotate("text", 
           x = label_x, y = label_y, 
           label = "PM", 
           size = 3.4, fontface = "bold", hjust = 0, color = "black")
fig4b

#####################
## Fig 4all
#####################

fig4all <- grid.arrange(fig4a, fig4b,                     
                        ncol = 2, nrow = 1,
                        layout_matrix = rbind(c(1,1,1,2,2,2,2)))

fig4all <- as_ggplot(fig4all) +                                
  draw_plot_label(label = c("A", "B"), size = 17,
                  x = c(0.01, 0.44), y = c(0.99, 0.99)) 

fig4all

ggsave("Fig4.pdf", plot=fig4all, path="~/Desktop/doctorate/ch2 mdd highway/manuscript_figures/", width = 10, height = 6, units="in", device = "pdf")
```

```{r, Table1}
#####################
## Build aggregated model
#####################
dengue_df_agg <- dengue_df_yearly_buffered
dengue_df_agg <- dengue_df_agg[which(as.Date(dengue_df_agg$year) > '2007-01-01'),]
dengue_df_agg$year_binary <- ifelse(as.Date(dengue_df_agg$year) > '2008-01-01',1,0)

leish_df_agg <- leish_df_yearly_buffered
leish_df_agg <- leish_df_agg[which(as.Date(leish_df_agg$year) > '2007-01-01'),]
leish_df_agg$year_binary <- ifelse(as.Date(leish_df_agg$year) > '2008-01-01',1,0)

# main specification aggregated effect estimate (for comparison)
dengue_yearly_model_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, 
                                  vcov = "cluster", 
                                  data = dengue_df_agg)

leish_yearly_model_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, 
                                  vcov = "cluster", 
                                  data = leish_df_agg)

#####################
## Table 1
#####################
table1 <- etable(dengue_yearly_model_agg,dengue_biannual_agg_model_dry,dengue_biannual_agg_model_rainy, 
                 leish_yearly_model_agg,leish_biannual_agg_model_dry,leish_biannual_agg_model_rainy, digits = 2)
table1 <- table1[c(3:8,9:12,16,15,14),2:7]
table1[13,] <- c(rep(65,6))
colnames(table1) <- c('Dengue Yearly', 'Dengue Biannual Dry', 'Dengue Biannual Rainy','Leish Yearly','Leish Biannual Dry', 'Leish Biannual Rainy')
rownames(table1) <- c("log(Urban)", "log(Ag)", "Sum Precip", "Mean Temp", "5km x Post-2008 Yearly", "5km x Post-2008 Biannual", "FEs", "Cluster", "Year", "Biannual", "Cor^2", "N", "Clusters")
table1 <- table1[c(5:6,1:4,7:13),]
table1
write.csv(table1, "~/Desktop/doctorate/ch2 mdd highway/manuscript_figures/table1.csv")
```
