==> 07_supplementary_figures.Rmd <==
---
title: "IOH and dengue supplementary figures"
author: "Alyson Singleton"
date: "`r Sys.Date()`"
output:
  html_document: default
  word_document: default
  pdf_document: default
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
# install packages, load libraries, set themes, knit formatting
pacman::p_load(tidyverse, dplyr, kableExtra, knitr, sf, raster, terra, spData, tmap, leaflet, ggplot2, spThin, mapdata, rnaturalearth, rnaturalearthdata, brazilmaps, devtools, brazilmaps, mapview, grid, gridExtra, RColorBrewer, raster, plm, fixest, cobalt, broom, ggpubr, gridExtra, cowplot, ggspatial, directlabels, scales, ggrepel)
theme_set(theme_bw())
knitr::opts_chunk$set(echo = F, results='asis', warning=F, message=F, cache=F,
                      fig.height=5, fig.width=10)

get_legend<-function(myggplot){
  tmp <- ggplot_gtable(ggplot_build(myggplot))
  leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
  legend <- tmp$grobs[[leg]]
  return(legend)
}

library(showtext)
font_add("Arial", "/Library/Fonts/Arial.ttf")  # Use the actual file path
showtext_auto()

```

```{r, load yearly case data}
vert_line_date <- as.Date('2008-01-01')

#for harper
# dengue_df_monthly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_monthly_full_dataset_c.csv")
# 
# dengue_df_monthly <- dengue_df_monthly %>%
#   group_by(tenkm,month) %>%
#   summarise(monthly_cases = sum(monthly_cases),
#             population = sum(population)) %>%
#   rename(group = tenkm)
# 
# write.csv(dengue_df_monthly,"~/Desktop/doctorate/ch2 mdd highway/data/harper_case_counts.csv")

###################
### dengue data ###
###################
### yearly data
dengue_df_yearly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_yearly_full_dataset_c.csv")
dengue_df_yearly$year <- as.factor(dengue_df_yearly$year)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_yearly$incidence <- (dengue_df_yearly$yearly_cases+1)/dengue_df_yearly$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_yearly$urban <- dengue_df_yearly$urban + 0.001
dengue_df_yearly$ag <- dengue_df_yearly$ag + 0.001

# double check complete cases
dengue_df_yearly <- dengue_df_yearly[complete.cases(dengue_df_yearly),]

# remove extreme outliers (all most likely due to inaccurate pop data)
dengue_df_yearly <- dengue_df_yearly[which(dengue_df_yearly$incidence < 0.5),]

# check how many are always zero
dengue_df_yearly_zeroes <- dengue_df_yearly %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(yearly_cases))
clusters_w_at_least_one_case <- dengue_df_yearly$cluster[which(dengue_df_yearly_zeroes$total_cases>0)]
dengue_df_yearly_at_least_one_case <- dengue_df_yearly[which(dengue_df_yearly$cluster %in% clusters_w_at_least_one_case),]

# remove PM
dengue_df_yearly_no_pm <- dengue_df_yearly[!(dengue_df_yearly$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_yearly_buffered <- dengue_df_yearly[which(dengue_df_yearly$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
dengue_df_yearly_buffered_no_pm <- dengue_df_yearly_buffered[!(dengue_df_yearly_buffered$cluster %in% 1),]

###################
### leish data ####
###################
### yearly data
leish_df_yearly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/leish_yearly_full_dataset_c.csv")

leish_df_yearly$year <- as.factor(leish_df_yearly$year)

# cases + 1 so 0's dont go to infinity when logged
leish_df_yearly$incidence <- (leish_df_yearly$yearly_cases+1)/leish_df_yearly$population

# land use vars + 0.001 so 0's dont go to infinity when logged
leish_df_yearly$urban <- leish_df_yearly$urban + 0.001
leish_df_yearly$ag <- leish_df_yearly$ag + 0.001

# double check complete cases
leish_df_yearly <- leish_df_yearly[complete.cases(leish_df_yearly),]

# remove extreme outliers
leish_df_yearly <- leish_df_yearly[which(leish_df_yearly$incidence < 0.5),]

# check how many are always zero
leish_df_yearly_zeroes <- leish_df_yearly  %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(yearly_cases))
leish_clusters_w_at_least_one_case <- leish_df_yearly$cluster[which(leish_df_yearly_zeroes$total_cases>0)]
leish_df_yearly_at_least_one_case <- leish_df_yearly[which(leish_df_yearly$cluster %in% leish_clusters_w_at_least_one_case),]

# remove PM
leish_df_yearly_no_pm <- leish_df_yearly[!(leish_df_yearly$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
leish_df_yearly_buffered <- leish_df_yearly[which(leish_df_yearly$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
leish_df_yearly_buffered_no_pm <- leish_df_yearly_buffered[!(leish_df_yearly_buffered$cluster %in% 1),]

##########################################
### dengue data confirmed and probable ###
##########################################
### yearly data
dengue_df_yearly_cp <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_yearly_full_dataset_cp.csv")
dengue_df_yearly_cp$year <- as.factor(dengue_df_yearly_cp$year)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_yearly_cp$incidence <- (dengue_df_yearly_cp$yearly_cases+1)/dengue_df_yearly_cp$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_yearly_cp$urban <- dengue_df_yearly_cp$urban + 0.001
dengue_df_yearly_cp$ag <- dengue_df_yearly_cp$ag + 0.001

# double check complete cases
dengue_df_yearly_cp <- dengue_df_yearly_cp[complete.cases(dengue_df_yearly_cp),]

# remove extreme outliers (all most likely due to inaccurate pop data)
dengue_df_yearly_cp <- dengue_df_yearly_cp[which(dengue_df_yearly_cp$incidence < 0.5),]

# check how many are always zero
dengue_df_yearly_zeroes <- dengue_df_yearly_cp %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(yearly_cases))
clusters_w_at_least_one_case <- dengue_df_yearly_cp$cluster[which(dengue_df_yearly_zeroes$total_cases>0)]
dengue_df_yearly_at_least_one_case <- dengue_df_yearly_cp[which(dengue_df_yearly_cp$cluster %in% clusters_w_at_least_one_case),]

# remove PM
dengue_df_yearly_cp_no_pm <- dengue_df_yearly_cp[!(dengue_df_yearly_cp$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_yearly_cp_buffered <- dengue_df_yearly_cp[which(dengue_df_yearly_cp$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
dengue_df_yearly_cp_buffered_no_pm <- dengue_df_yearly_cp_buffered[!(dengue_df_yearly_cp_buffered$cluster %in% 1),]

###############################################################################
### dengue data only clusters w at least one disease case reported pre/post ###
###############################################################################
### yearly data
dengue_df_yearly_prepost <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km_prepost_req/dengue_yearly_full_dataset_c.csv")
dengue_df_yearly_prepost$year <- as.factor(dengue_df_yearly_prepost$year)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_yearly_prepost$incidence <- (dengue_df_yearly_prepost$yearly_cases+1)/dengue_df_yearly_prepost$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_yearly_prepost$urban <- dengue_df_yearly_prepost$urban + 0.001
dengue_df_yearly_prepost$ag <- dengue_df_yearly_prepost$ag + 0.001

# double check complete cases
dengue_df_yearly_prepost <- dengue_df_yearly_prepost[complete.cases(dengue_df_yearly_prepost),]

# remove extreme outliers (all most likely due to inaccurate pop data)
dengue_df_yearly_prepost <- dengue_df_yearly_prepost[which(dengue_df_yearly_prepost$incidence < 0.5),]

# check how many are always zero
dengue_df_yearly_zeroes <- dengue_df_yearly_prepost %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(yearly_cases))
clusters_w_at_least_one_case <- dengue_df_yearly_prepost$cluster[which(dengue_df_yearly_zeroes$total_cases>0)]
dengue_df_yearly_at_least_one_case <- dengue_df_yearly_prepost[which(dengue_df_yearly_prepost$cluster %in% clusters_w_at_least_one_case),]

# remove PM
dengue_df_yearly_prepost_no_pm <- dengue_df_yearly_prepost[!(dengue_df_yearly_prepost$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_yearly_prepost_buffered <- dengue_df_yearly_prepost[which(dengue_df_yearly_prepost$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
dengue_df_yearly_prepost_buffered_no_pm <- dengue_df_yearly_prepost_buffered[!(dengue_df_yearly_prepost_buffered$cluster %in% 1),]
```

```{r, load biannaul case data}
vert_line_date <- as.Date('2008-01-01')

###################
### dengue data ###
###################
### biannual data
dengue_df_biannual <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_biannual_full_dataset_c.csv")
dengue_df_biannual$biannual_date <- as.factor(dengue_df_biannual$biannual_date)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_biannual$incidence <- (dengue_df_biannual$biannual_cases+1)/dengue_df_biannual$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_biannual$urban <- dengue_df_biannual$urban + 0.001
dengue_df_biannual$ag <- dengue_df_biannual$ag + 0.001

# double check complete cases
dengue_df_biannual <- dengue_df_biannual[complete.cases(dengue_df_biannual),]

# remove extreme outliers
dengue_df_biannual <- dengue_df_biannual[which(dengue_df_biannual$incidence < 0.5),]

#check how many are always zero
dengue_df_biannual_zeroes <- dengue_df_biannual %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(biannual_cases))
clusters_w_at_least_one_case_biannual <- dengue_df_biannual$cluster[which(dengue_df_biannual_zeroes$total_cases>0)]
dengue_df_biannaul_at_least_one_case <- dengue_df_biannual[which(dengue_df_biannual$cluster %in% clusters_w_at_least_one_case_biannual),]

#remove PM
dengue_df_biannual_no_pm <- dengue_df_biannual[!(dengue_df_biannual$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_biannual_buffered <- dengue_df_biannual[which(dengue_df_biannual$all_cutoffs %in% c(1,2,4,5,6,7,0)),]

###################
### leish data ####
###################

### biannaul data
leish_df_biannual <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/leish_biannual_full_dataset_c.csv")
leish_df_biannual$biannual_date <- as.factor(leish_df_biannual$biannual_date)

# cases + 1 so 0's dont go to infinity when logged
leish_df_biannual$incidence <- (leish_df_biannual$biannual_cases+1)/leish_df_biannual$population

# land use vars + 0.001 so 0's dont go to infinity when logged
leish_df_biannual$urban <- leish_df_biannual$urban + 0.001
leish_df_biannual$ag <- leish_df_biannual$ag + 0.001

# double check complete cases
leish_df_biannual <- leish_df_biannual[complete.cases(leish_df_biannual),]

# remove extreme outliers
leish_df_biannual <- leish_df_biannual[which(leish_df_biannual$incidence < 0.5),]

# check how many are always zero
leish_df_biannual_zeroes <- leish_df_biannual  %>% 
  group_by(cluster) %>% 
  summarize(total_cases = sum(biannual_cases))
leish_clusters_w_at_least_one_case_biannual <- leish_df_biannual$cluster[which(leish_df_biannual_zeroes$total_cases>0)]
leish_df_biannual_at_least_one_case <- leish_df_biannual[which(leish_df_biannual$cluster %in% leish_clusters_w_at_least_one_case_biannual),]

# remove PM
leish_df_biannual_no_pm <- leish_df_biannual[!(leish_df_biannual$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
leish_df_biannual_buffered <- leish_df_biannual[which(leish_df_biannual$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
```

```{r, STable1}
dengue_yearly_model <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffered)
leish_yearly_model <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + sum_precip + mean_temp + log(urban) + log(ag) | cluster + year, vcov = "cluster", data = leish_df_yearly_buffered)

#####################
## STable 1
#####################
stable1 <- etable(dengue_yearly_model, leish_yearly_model, digits = 2)
stable1 <- stable1[c(3:10,37,11:31,35,34,36),2:3]
stable1[33,] <- c(rep(65,2))
stable1[9,] <- c("----","----")
colnames(stable1) <- c('Dengue Yearly', 'Leish Yearly')
rownames(stable1) <- c("5km x 2000", "5km x 2001", "5km x 2002", "5km x 2003", "5km x 2004", "5km x 2005", "5km x 2006", "5km x 2007", "5km x 2008", "5km x 2009", "5km x 2010", "5km x 2011", "5km x 2012", "5km x 2013", "5km x 2014", "5km x 2015", "5km x 2016", "5km x 2017", "5km x 2018", "5km x 2019", "5km x 2020", "5km x 2021", "5km x 2022", "log(Urban)", "log(Ag)", "Sum Precip", "Mean Temp", "FEs", "Cluster", "Year", "Cor^2", "N", "Clusters")
stable1
write.csv(stable1, "~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/stable1.csv")
```

```{r, STable2}
#####################
## Build aggregated models
#####################
dengue_df_agg <- dengue_df_yearly
dengue_df_agg <- dengue_df_agg[which(as.Date(dengue_df_agg$year) > '2007-01-01'),]
dengue_df_agg$year_binary <- ifelse(as.Date(dengue_df_agg$year) > '2008-01-01',1,0)

dengue_df_agg_buffered <- dengue_df_yearly_buffered
dengue_df_agg_buffered <- dengue_df_agg_buffered[which(as.Date(dengue_df_agg_buffered$year) > '2007-01-01'),]
dengue_df_agg_buffered$year_binary <- ifelse(as.Date(dengue_df_agg_buffered$year) > '2008-01-01',1,0)

dengue_df_agg_buffered_cp <- dengue_df_yearly_cp_buffered
dengue_df_agg_buffered_cp <- dengue_df_agg_buffered_cp[which(as.Date(dengue_df_agg_buffered_cp$year) > '2007-01-01'),]
dengue_df_agg_buffered_cp$year_binary <- ifelse(as.Date(dengue_df_agg_buffered_cp$year) > '2008-01-01',1,0)

dengue_df_agg_buffered_prepost <- dengue_df_yearly_prepost_buffered
dengue_df_agg_buffered_prepost <- dengue_df_agg_buffered_prepost[which(as.Date(dengue_df_agg_buffered_prepost$year) > '2007-01-01'),]
dengue_df_agg_buffered_prepost$year_binary <- ifelse(as.Date(dengue_df_agg_buffered_prepost$year) > '2008-01-01',1,0)

## main specification
dengue_yearly_model_main_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg_buffered)

## population weighting
dengue_df_agg_buffered_no_pm <- dengue_df_agg_buffered[!(dengue_df_agg_buffered$cluster %in% 1),]
dengue_yearly_model_pop_weight_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg_buffered_no_pm, weights=~population)

## no Puerto Maldonado
dengue_yearly_model_no_PM_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg_buffered_no_pm)

## changing the spatial boundary delineating treatment and control groups
dengue_yearly_model_onekm_agg <- fepois(incidence ~ year_binary*onekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg)
dengue_yearly_model_tenkm_agg <- fepois(incidence ~ year_binary*tenkm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg)

## changing the buffer zone size between treatment and control groups
dengue_yearly_model_no_buffer_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg)
dengue_df_yearly_buffer_bigger_agg <- dengue_df_agg[which(dengue_df_agg$all_cutoffs %in% c(1,2,5,6,7,0)),]
dengue_yearly_model_buffer_bigger_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffer_bigger_agg)

## confirmed cases only v confirmed and probable cases
dengue_yearly_model_confirmed_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg_buffered_cp)

## only include clusters where at least one disease case reported both pre and post treatment
dengue_yearly_model_prepost_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_agg_buffered_prepost)

#####################
## STable S2
#####################
stable2 <- etable(dengue_yearly_model_main_agg,dengue_yearly_model_no_PM_agg,dengue_yearly_model_pop_weight_agg,dengue_yearly_model_onekm_agg,
                  dengue_yearly_model_tenkm_agg,dengue_yearly_model_no_buffer_agg,dengue_yearly_model_buffer_bigger_agg,dengue_yearly_model_confirmed_agg, dengue_yearly_model_prepost_agg, digits = 2)
stable2 <- stable2[c(3:9,16,15,14),2:10]
stable2[10,] <- (as.numeric(gsub(",","",stable2[9,]))+1)/15
stable2[10,9] <- 49
colnames(stable2) <- c('Main model', 'No PM', 'Pop weight no PM','1km boundary','10km boundary','No buffer (<5km v >5km)', 'Large buffer (<5km v >20km)', 'Conf & prob cases', 'Units reporting pre & post')
rownames(stable2) <- c("log(Urban)", "log(Ag)", "Sum Precip", "Mean Temp", "5km x Post-2008", "1km x Post-2008", "10km x Post-2008", "Cor^2", "N", "Clusters")
stable2 <- stable2[c(5:7,1:4, 8:10),]
stable2
write.csv(stable2, "~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/stable2.csv")
```

```{r, STable3}
#####################
## Build aggregated datasets
#####################
dengue_df_agg <- dengue_df_yearly_buffered
dengue_df_agg <- dengue_df_agg[which(as.Date(dengue_df_agg$year) > '2007-01-01'),]
dengue_df_agg$year_binary <- ifelse(as.Date(dengue_df_agg$year) > '2008-01-01',1,0)
dengue_df_agg_biannual <- dengue_df_biannual_buffered
dengue_df_agg_biannual <- dengue_df_agg_biannual[which(as.Date(dengue_df_biannual_buffered$biannual_date) > '2006-10-01'),]
dengue_df_agg_biannual$biannual_binary <- ifelse(as.Date(dengue_df_agg_biannual$biannual_date) > '2008-04-01',1, 0)
dengue_df_agg_biannual$month <- format(as.Date(dengue_df_agg_biannual$biannual_date), "%m")
dengue_df_agg_biannual_dry <- dengue_df_agg_biannual[which(dengue_df_agg_biannual$month == "04"),]
dengue_df_agg_biannual_rainy <- dengue_df_agg_biannual[which(dengue_df_agg_biannual$month == "10"),]

leish_df_agg <- leish_df_yearly_buffered
leish_df_agg <- leish_df_agg[which(as.Date(leish_df_agg$year) > '2007-01-01'),]
leish_df_agg$year_binary <- ifelse(as.Date(leish_df_agg$year) > '2008-01-01',1,0)
leish_df_agg_biannual <- leish_df_biannual_buffered
leish_df_agg_biannual <- leish_df_agg_biannual[which(as.Date(leish_df_biannual_buffered$biannual_date) > '2006-10-01'),]
leish_df_agg_biannual$biannual_binary <- ifelse(as.Date(leish_df_agg_biannual$biannual_date) > '2008-04-01',1, 0)
leish_df_agg_biannual$month <- format(as.Date(leish_df_agg_biannual$biannual_date), "%m")
leish_df_agg_biannual_dry <- leish_df_agg_biannual[which(leish_df_agg_biannual$month == "04"),]
leish_df_agg_biannual_rainy <- leish_df_agg_biannual[which(leish_df_agg_biannual$month == "10"),]

#####################
## Build aggregated models
#####################
dengue_yearly_model_agg_quad <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip^2 + mean_temp^2 | cluster + year, 
                                  vcov = "cluster", 
                                  data = dengue_df_agg)
dengue_biannual_agg_model_dry_quad <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip^2 + mean_temp^2 | cluster + biannual_date, vcov = "cluster", data = dengue_df_agg_biannual_dry)
dengue_biannual_agg_model_rainy_quad <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip^2 + mean_temp^2 | cluster + biannual_date, vcov = "cluster", data = dengue_df_agg_biannual_rainy)

leish_yearly_model_agg_quad <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip^2 + mean_temp^2 | cluster + year, 
                                  vcov = "cluster", 
                                  data = leish_df_agg)

leish_biannual_agg_model_dry_quad <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip^2 + mean_temp^2 | cluster + biannual_date, vcov = "cluster", data = leish_df_agg_biannual_dry)
leish_biannual_agg_model_rainy_quad <- fepois(incidence ~ biannual_binary*fivekm + log(urban) + log(ag) + sum_precip^2 + mean_temp^2 | cluster + biannual_date, vcov = "cluster", data = leish_df_agg_biannual_rainy)

#####################
## STable 3
#####################
stable3 <- etable(dengue_yearly_model_agg_quad,dengue_biannual_agg_model_dry_quad,dengue_biannual_agg_model_rainy_quad, 
                 leish_yearly_model_agg_quad,leish_biannual_agg_model_dry_quad,leish_biannual_agg_model_rainy_quad, digits = 2)
stable3 <- stable3[c(3:8,9:12,16,15,14),2:7]
stable3[13,] <- c(rep(65,6))
colnames(stable3) <- c('Dengue Yearly', 'Dengue Biannual Dry', 'Dengue Biannual Rainy','Leish Yearly','Leish Biannual Dry', 'Leish Biannual Rainy')
rownames(stable3) <- c("log(Urban)", "log(Ag)", "Sum Precip^2", "Mean Temp^2", "5km x Post-2008 Yearly", "5km x Post-2008 Biannual", "FEs", "Cluster", "Year", "Biannual", "Cor^2", "N", "Clusters")
stable3 <- stable3[c(5:6,1:4,7:13),]
stable3
write.csv(stable3, "~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/stable3.csv")
```

```{r, SFig S1}
#####################
## SFig 1a
#####################
dengue_df_yearly$case_yn <- ifelse(dengue_df_yearly$yearly_cases>0, 1, 0)
dengue_coverage_yearly <- dengue_df_yearly %>%
  group_by(year, fivekm) %>%
  summarize(coverage = sum(case_yn))

## to determine dengue distribution before paving
# dengue_coverage_yearly <- dengue_df_yearly %>%
#   group_by(year, cluster) %>%
#   summarize(coverage = sum(case_yn),
#             fivekm = max(fivekm))

dengue_coverage_yearly$fivekm <- as.character(dengue_coverage_yearly$fivekm)
dengue_coverage_yearly <- dengue_coverage_yearly[0:42,]
dengue_coverage_yearly$year <- format(as.Date(dengue_coverage_yearly$year, format="%Y-%m-%d"),"%Y")

sfig1a <- ggplot(dengue_coverage_yearly) +
  geom_col(aes(x= year, y=coverage, fill=fivekm), width=0.5, color="white", position = position_dodge(width = 0.5)) +
  scale_fill_manual(name = "", values=c('#648FFF', '#E04490'), labels=c('>5km', '<5km'),) +
  geom_vline(aes(xintercept=("2008")), linetype='dashed', size=0.4) +
  xlab("") + ylab("# units w/\ndengue\ncase") + 
  theme_bw()+
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=10),
        axis.text.x=element_text(size=10, angle=45, vjust=0.45),
        axis.text = element_text(size=16),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "bottom",
        strip.text.x = element_text(size = 14))
sfig1a
sfig1_legend <- get_legend(sfig1a)
sfig1a <- sfig1a + theme(legend.position = "none")

#####################
## SFig 1b
#####################
leish_df_yearly$case_yn <- ifelse(leish_df_yearly$yearly_cases>0, 1, 0)
leish_coverage_yearly <- leish_df_yearly %>%
  group_by(year, fivekm) %>%
  summarize(coverage = sum(case_yn))

leish_coverage_yearly$fivekm <- as.character(leish_coverage_yearly$fivekm)
leish_coverage_yearly <- leish_coverage_yearly[0:42,]
leish_coverage_yearly$year <- format(as.Date(leish_coverage_yearly$year, format="%Y-%m-%d"),"%Y")

sfig1b <- ggplot(leish_coverage_yearly) +
  geom_col(aes(x= year, y=coverage, fill=fivekm), color="white", width=0.5, position = position_dodge(width = 0.5)) +
  scale_fill_manual(name = "Treatment", values=c('#648FFF', '#E04490'), labels=c('>5km', '<5km'),) +
  geom_vline(aes(xintercept=("2008")), linetype='dashed', size=0.4) +
  xlab("Year") + ylab("# units w/\nleish\ncase") + 
  theme_bw() +
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=12, angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=10, angle=45, vjust=0.45),
        axis.text = element_text(size=16),
        legend.text=element_text(size=14),
        legend.title=element_text(size=14),
        legend.position = "none",
        strip.text.x = element_text(size = 14))
sfig1b

#####################
## SFig 1all
#####################

sfig1all <- grid.arrange(sfig1a, sfig1b, sfig1_legend,
                         ncol = 1, nrow = 3,
                         layout_matrix = rbind(c(1),c(2),c(3)), 
                         heights=c(3,3,1))

sfig1all <- as_ggplot(sfig1all) +                                
  draw_plot_label(label = c("A", "B"), size = 14,
                  x = c(0.05, 0.05), y = c(1, 0.58)) 
sfig1all

ggsave("SFig1.pdf", plot=sfig1all, path="~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 8, height = 7, units="in", device = "pdf")

```

```{r, SFig S2}
## SFig S2 in 06_sfig_s2.R file
```

```{r, SFig S3}
#####################
## SFig 3
#####################

## main specification
dengue_yearly_model_main <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffered)

## population weighting
dengue_yearly_model_pop_weight <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffered_no_pm, weights=~population)

## no Puerto Maldonado
dengue_yearly_model_no_PM <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffered_no_pm)

## changing the spatial boundary delineating treatment and control groups
dengue_yearly_model_onekm <- fepois(incidence ~ i(year, onekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly)
dengue_yearly_model_tenkm <- fepois(incidence ~ i(year, tenkm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly)

## changing the buffer zone size between treatment and control groups
dengue_df_yearly_no_buffer <- dengue_df_yearly[which(dengue_df_yearly$all_cutoffs %in% c(1,2,3,4,5,6,7,0)),]
dengue_yearly_model_no_buffer <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_no_buffer)
dengue_df_yearly_buffer_bigger <- dengue_df_yearly[which(dengue_df_yearly$all_cutoffs %in% c(1,2,5,6,7,0)),]
dengue_yearly_model_buffer_bigger <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffer_bigger)

## confirmed cases only v confirmed and probable cases
dengue_yearly_model_confirmed <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_cp_buffered)

## only include clusters where at least one disease case reported both pre and post treatment
dengue_yearly_model_prepost <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_prepost_buffered)

df <- rbind(
  as.data.frame(cbind(dengue_yearly_model_no_PM$coeftable[1:22,],c(rep("No PM",22)))),
  as.data.frame(cbind(dengue_yearly_model_pop_weight$coeftable[1:22,],c(rep("Pop weight no PM",22)))),
  as.data.frame(cbind(dengue_yearly_model_onekm$coeftable[1:22,],c(rep("1km",22)))),
  as.data.frame(cbind(dengue_yearly_model_tenkm$coeftable[1:22,],c(rep("10km",22)))),
  as.data.frame(cbind(dengue_yearly_model_no_buffer$coeftable[1:22,],c(rep("No buffer",22)))),
  as.data.frame(cbind(dengue_yearly_model_buffer_bigger$coeftable[1:22,],c(rep("Big buffer",22)))),
  as.data.frame(cbind(dengue_yearly_model_confirmed$coeftable[1:22,],c(rep("Confirmed cases",22)))),
  as.data.frame(cbind(dengue_yearly_model_prepost$coeftable[1:22,],c(rep("Pre post",22)))),
  as.data.frame(cbind(dengue_yearly_model_main$coeftable[1:22,],c(rep("Main",22)))))

colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value', 'Model')
df$year <- c(seq(as.Date("2000-01-01"), as.Date("2007-01-01"), by="year"),
             seq(as.Date("2009-01-01"), as.Date("2022-01-01"), by="year"))
df$estimate <- as.numeric(df$estimate)
df$std_error <- as.numeric(df$std_error)
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
df$Model <- factor(as.character(df$Model), levels=c('No PM', 'Pop weight no PM','1km','10km','No buffer', 'Big buffer', 'Confirmed cases', "Pre post", 'Main'))

#consider updating this so the main model is just bold black ouline with no fill in the point estimate
sfig3a <- ggplot(df, aes(group=Model)) +
  geom_hline(aes(yintercept=0), colour='red', linewidth=.4) +
  geom_errorbar(aes(x=year, ymax=upper, ymin=lower, color=Model), width=0.1, linewidth=0.5, alpha=0.8) +
  geom_vline(aes(xintercept=as.Date("2008-01-01")), linetype='dashed', linewidth=0.4) +
  geom_point(aes(x=as.Date("2008-01-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(year, estimate, fill=Model), color='black', size=3, shape=21, alpha=0.6) +
  scale_color_manual(name = "Model specification", values=list('No PM'='#DC267F','Pop weight no PM'="#FE6100",'1km'='#FFB000','10km'='#FFE518','No buffer'='#85CE76', 'Big buffer'='#126E29','Confirmed cases'="#648FFF", 'Pre post'="#785EF0", 'Main' = "#2b175e"),
                     labels=list('No PM', 'Pop weight no PM','1km boundary','10km boundary','No buffer (<5km v >5km)', 'Large buffer (<5km v >20km)', 'Conf & prob cases', 'Units reporting pre & post', 'Main model'))+
  scale_fill_manual(name = "Model specification", values=list('No PM'='#DC267F','Pop weight no PM'="#FE6100",'1km'='#FFB000','10km'='#FFE518','No buffer'='#85CE76', 'Big buffer'='#126E29','Confirmed cases'="#648FFF", 'Pre post'="#785EF0", 'Main' = NA),
                    labels=list('No PM', 'Pop weight no PM','1km boundary','10km boundary','No buffer (<5km v >5km)', 'Large buffer (<5km v >20km)', 'Conf & prob cases', 'Units reporting pre & post', 'Main model'))+
  xlab("Year") + ylab("Percent\nchange\ndengue\nincidence") + 
  theme_bw()+
  scale_y_continuous(labels = scales::percent) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=18),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=12),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "right",
        strip.text.x = element_text(size = 12))
sfig3a
sfig3_legend <- get_legend(sfig3a)
sfig3a <- sfig3a + theme(legend.position = "none")

df_limits <- df
df_limits$upper <- ifelse(df$upper>8, 7.99, df$upper)

# df_limits_test <- df_limits %>% 
#   filter(Model %in% c('Main','10km', 'No buffer'))

sfig3b <- ggplot(df_limits, aes(group=Model)) +
  geom_hline(aes(yintercept=0), colour='red', linewidth=.4) +
  geom_errorbar(aes(x=year, ymax=upper, ymin=lower, color=Model), width=0.1, linewidth=0.5, alpha=0.8) +
  geom_vline(aes(xintercept=as.Date("2008-01-01")), linetype='dashed', linewidth=0.4) +
  geom_point(aes(x=as.Date("2008-01-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(year, estimate, fill=Model), color='black', size=3, shape=21, alpha=0.6) +
  scale_color_manual(name = "Model specification", values=list('No PM'='#DC267F','Pop weight no PM'="#FE6100",'1km'='#FFB000','10km'='#FFE518','No buffer'='#85CE76', 'Big buffer'='#126E29','Confirmed cases'="#648FFF", 'Pre post'="#785EF0", 'Main' = "#2b175e"),
                     labels=list('No PM', 'Pop weight no PM','1km boundary','10km boundary','No buffer (<5km v >5km)', 'Large buffer (<5km v >20km)', 'Conf & prob cases', 'Units reporting pre & post', 'Main model'))+
  scale_fill_manual(name = "Model specification", values=list('No PM'='#DC267F','Pop weight no PM'="#FE6100",'1km'='#FFB000','10km'='#FFE518','No buffer'='#85CE76', 'Big buffer'='#126E29','Confirmed cases'="#648FFF", 'Pre post'="#785EF0", 'Main' = "#2b175e"),
                    labels=list('No PM', 'Pop weight no PM','1km boundary','10km boundary','No buffer (<5km v >5km)', 'Large buffer (<5km v >20km)', 'Conf & prob cases', 'Units reporting pre & post', 'Main model'))+
  xlab("Year") + ylab("Percent\nchange\ndengue\nincidence") + 
  theme_bw()+
  scale_y_continuous(labels = scales::percent, limits=c(-1,8)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=18),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=12),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=12),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "none",
        strip.text.x = element_text(size = 12))
sfig3b



#####################
## SFig 3all
#####################

sfig3all <- grid.arrange(sfig3a, sfig3b, sfig3_legend,
                         ncol = 4, nrow = 2,
                         layout_matrix = rbind(c(1,1,1,3),c(2,2,2,3)), 
                         heights=c(1,1))

sfig3all <- as_ggplot(sfig3all) +                                
  draw_plot_label(label = c("A", "B"), size = 14,
                  x = c(0.111, 0.111), y = c(0.99, 0.49)) 
sfig3all

ggsave("SFig3.pdf", plot=sfig3all, "~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 12, height = 10, units = c("in"), device="pdf")

```

```{r, SFig S4}
#####################
## SFig 4
#####################
# Permutation Inference Analysis

### aggregated data (treatment)
dengue_df_agg <- dengue_df_yearly_buffered
dengue_df_agg <- dengue_df_agg[which(as.Date(dengue_df_agg$year) > '2007-01-01'),]
dengue_df_agg$year_binary <- ifelse(as.Date(dengue_df_agg$year) > '2008-01-01',1,0)

# main specification aggregated effect estimate (for comparison)
dengue_yearly_model_agg <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, 
                                  vcov = "cluster", 
                                  data = dengue_df_agg)
dengue_yearly_model_agg_effect_est <- coeftable(dengue_yearly_model_agg)[5,1]
dengue_yearly_model_agg_effect_est <- exp(dengue_yearly_model_agg_effect_est) - 1
  
#####################
## SFig 4a
#####################

# Full (randomize treatment across clusters and years)
permuted_effects_stor_full <- c()
for(i in 1:1000){
  set.seed(i)
  dengue_df_yearly_permuted_full  <- transform(dengue_df_agg, fivekm = sample(fivekm))
  dengue_yearly_model_permuted_full <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, 
                                              vcov = "cluster", 
                                              data = dengue_df_yearly_permuted_full)
  permuted_effect_est_full <- coeftable(dengue_yearly_model_permuted_full)[6,1]
  permuted_effects_stor_full <- c(permuted_effects_stor_full, permuted_effect_est_full)
}

permuted_effects_stor_full <- exp(permuted_effects_stor_full) - 1
sfig4a <- ggplot() +
  geom_histogram(aes(permuted_effects_stor_full), binwidth = 0.03, fill="grey") +
  geom_vline(xintercept=dengue_yearly_model_agg_effect_est, color='red') +
  xlab("Bootstrapped effect estimate") + ylab("") + 
  theme_bw()+
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=10,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=10),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=16),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 14))
sfig4a

#####################
## SFig 4b
#####################

# Block (randomize treatment across clusters but maintain temporal structure)
# run through each cluster and randomly assign in or out of treatment (using derived probabilities)
dengue_df_yearly_permuted_block <- dengue_df_agg
permuted_effects_stor_block <- c()
for(i in 1:1000){
  set.seed(i)
  prob_1 <- length(which(dengue_df_agg$fivekm==1))/length(dengue_df_agg$fivekm)
  #randomly assign treatment to each spatial unit
  for(k in 1:length(unique(dengue_df_agg$cluster))){
    random_treatment <- rbinom(1, 1, prob_1)
    dengue_df_yearly_permuted_block$fivekm[which(dengue_df_yearly_permuted_block$cluster==k)] <- if (random_treatment==1) {rep(1,15)} else {rep(0,15)}
  }
  dengue_yearly_model_permuted_block <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, 
                                               vcov = "cluster", 
                                               data = dengue_df_yearly_permuted_block)
  permuted_effect_est_block <- coeftable(dengue_yearly_model_permuted_block)[5,1]
  permuted_effects_stor_block <- c(permuted_effects_stor_block, permuted_effect_est_block)
}

permuted_effects_stor_block <- exp(permuted_effects_stor_block) - 1
sfig4b <- ggplot() +
  geom_histogram(aes(permuted_effects_stor_block), binwidth = 0.03, fill="grey") +
  geom_vline(xintercept=dengue_yearly_model_agg_effect_est, color='red') +
  xlab("Bootstrapped effect estimate") + ylab("") + 
  theme_bw()+
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=10,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=10),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=16),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 14))
sfig4b

#####################
## SFig 4c
#####################

# Within (randomly assign the years that are flagged as treated within the districts that are truly part of the treatment group)
dengue_df_yearly_permuted_within <- dengue_df_agg
permuted_effects_stor_within <- c()
for(i in 1:1000){
  set.seed(i)
  #randomly assign treatment to each spatial unit
  for(k in 1:length(unique(dengue_df_agg$cluster))){
    #random_treatment <- rbinom(1, 1, prob_1)
    dengue_df_yearly_permuted_within$year_binary[which(dengue_df_yearly_permuted_within$cluster==k)] <- sample(c(0,rep(1,14)))
  }
  dengue_yearly_model_permuted_within <- fepois(incidence ~ year_binary*fivekm + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, 
                                               vcov = "cluster", 
                                               data = dengue_df_yearly_permuted_within)
  permuted_effect_est_within <- coeftable(dengue_yearly_model_permuted_within)[5,1]
  permuted_effects_stor_within <- c(permuted_effects_stor_within, permuted_effect_est_within)
}

permuted_effects_stor_within <- exp(permuted_effects_stor_within) - 1
sfig4c <- ggplot() +
  geom_histogram(aes(permuted_effects_stor_within), binwidth = 0.03, fill="grey") +
  geom_vline(xintercept=dengue_yearly_model_agg_effect_est, color='red') +
  xlab("Bootstrapped effect estimate") + ylab("") + 
  theme_bw()+
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=10,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=10),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=16),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 14))
sfig4c

#####################
## SFig 4all
#####################

sfig4all <- grid.arrange(sfig4a, sfig4b, sfig4c,
                         nrow = 1)

sfig4all <- as_ggplot(sfig4all) +                                
  draw_plot_label(label = c("A", "B","C"), size = 13,
                  x = c(0.00, 0.335, 0.675), y = c(1, 1, 1)) 
sfig4all

ggsave("SFig4.pdf", plot=sfig4all, path="~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 12, height = 5, units="in", device = "pdf")

```

```{r, SFig S5}
#####################
## SFig 5a
#####################
#population
dengue_df_yearly_pop <- dengue_df_yearly_buffered %>%
  group_by(year,fivekm) %>%
  summarise(pop_sum = sum(population))

dengue_df_yearly_pop_no_pm <- dengue_df_yearly_buffered_no_pm %>%
  group_by(year,fivekm) %>%
  summarise(pop_sum = sum(population))
dengue_df_yearly_pop_no_pm$fivekm <- ifelse(dengue_df_yearly_pop_no_pm$fivekm==1,2,0)
dengue_df_yearly_pop_no_pm <- dengue_df_yearly_pop_no_pm[which(dengue_df_yearly_pop_no_pm$fivekm==2),]

dengue_df_yearly_pop <- rbind(dengue_df_yearly_pop, dengue_df_yearly_pop_no_pm)

dengue_df_yearly_pop$fivekm <- as.character(dengue_df_yearly_pop$fivekm)
dengue_df_yearly_pop$year <- as.Date(dengue_df_yearly_pop$year)

sfig5a <- ggplot(dengue_df_yearly_pop) +
  geom_line(aes(x=year, y=pop_sum, color=fivekm)) +
  geom_vline(xintercept=vert_line_date,linetype='dashed') +
  ggtitle("Population") +
  xlab("Year") + ylab("") + 
  scale_color_manual(name = "Exposure", values=c("#648FFF","#E04490", "#FFD218"), labels=c('Far (>10km)', 'Near (<5km)', 'Near no PM (<5km)'),) +
  scale_y_continuous(labels = label_number(suffix = "k", scale = 1e-3)) +
  theme_bw()+
  theme(plot.title = element_text(size=12, face="bold"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=14),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=10),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "none",
        strip.text.x = element_text(size = 10))
sfig5a

#####################
## SFig 5b
#####################
#urban area
dengue_df_yearly_urban <- dengue_df_yearly_buffered %>%
  group_by(year,fivekm) %>%
  summarise(urban_sum = sum(urban))

dengue_df_yearly_urban_no_pm <- dengue_df_yearly_buffered_no_pm %>%
  group_by(year,fivekm) %>%
  summarise(urban_sum = sum(urban))
dengue_df_yearly_urban_no_pm$fivekm <- ifelse(dengue_df_yearly_urban_no_pm$fivekm==1,2,0)
dengue_df_yearly_urban_no_pm <- dengue_df_yearly_urban_no_pm[which(dengue_df_yearly_urban_no_pm$fivekm==2),]

dengue_df_yearly_urban <- rbind(dengue_df_yearly_urban, dengue_df_yearly_urban_no_pm)

dengue_df_yearly_urban$fivekm <- as.character(dengue_df_yearly_urban$fivekm)
dengue_df_yearly_urban$year <- as.Date(dengue_df_yearly_urban$year)

sfig5b <- ggplot(dengue_df_yearly_urban) +
  geom_line(aes(x=year, y=urban_sum, color=fivekm)) +
  geom_vline(xintercept=vert_line_date,linetype='dashed') +
  ggtitle("Urban Area (m^2)") +
  xlab("Year") + ylab("") + 
  scale_color_manual(name = "Exposure", values=c("#648FFF","#E04490", "#FFD218"), labels=c('Far (>10km)', 'Near (<5km)', 'Near no PM (<5km)'),) +
  theme_bw()+
  theme(plot.title = element_text(size=12, face="bold"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=14),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=13),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=10),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "none",
        strip.text.x = element_text(size = 10))
sfig5b

#####################
## SFig 5c
#####################
#agricultural area
dengue_df_yearly_ag <- dengue_df_yearly_buffered %>%
  group_by(year,fivekm) %>%
  summarise(ag_sum = sum(ag))

dengue_df_yearly_ag_no_pm <- dengue_df_yearly_buffered_no_pm %>%
  group_by(year,fivekm) %>%
  summarise(ag_sum = sum(ag))
dengue_df_yearly_ag_no_pm$fivekm <- ifelse(dengue_df_yearly_ag_no_pm$fivekm==1,2,0)
dengue_df_yearly_ag_no_pm <- dengue_df_yearly_ag_no_pm[which(dengue_df_yearly_ag_no_pm$fivekm==2),]

dengue_df_yearly_ag <- rbind(dengue_df_yearly_ag, dengue_df_yearly_ag_no_pm)

dengue_df_yearly_ag$fivekm <- as.character(dengue_df_yearly_ag$fivekm)
dengue_df_yearly_ag$year <- as.Date(dengue_df_yearly_ag$year)

sfig5c <- ggplot(dengue_df_yearly_ag) +
  geom_line(aes(x=year, y=ag_sum, color=fivekm)) +
  geom_vline(xintercept=vert_line_date,linetype='dashed') +
  ggtitle("Agricultural Area (m^2)") +
  xlab("Year") + ylab("") + 
  scale_color_manual(name = "Exposure", values=c("#648FFF","#E04490", "#FFD218"), labels=c('Far (>10km)', 'Near (<5km)', 'Near no PM (<5km)'),) +
  theme_bw()+
  theme(plot.title = element_text(size=12, face="bold"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=14),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=10),
        axis.text = element_text(size=10),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "bottom",
        strip.text.x = element_text(size = 10))
sfig5c
sfig5_legend <- get_legend(sfig5c)
sfig5c <- sfig5c + theme(legend.position = "none")

#####################
## SFig 5all
#####################

sfig5all <- grid.arrange(sfig5a, sfig5b, sfig5c, sfig5_legend,
                         ncol = 1, nrow = 4,
                         layout_matrix = rbind(c(1),c(2),c(3),c(4)), 
                         heights=c(4,4,4,1))

sfig5all <- as_ggplot(sfig5all) +                                
  draw_plot_label(label = c("A", "B", "C"), size = 14,
                  x = c(0.01, 0.01, 0.01), y = c(1, 0.692, 0.385)) 
sfig5all

ggsave("SFig5.pdf", plot=sfig5all, path="~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 6, height = 10, units="in", device = "pdf")

```

```{r, SFig S6}
### biannual model
dengue_biannual_model <- fepois(incidence ~ i(biannual_date, fivekm, ref = '2008-04-01') + sum_precip + mean_temp + log(urban) + log(ag) | cluster + biannual_date , vcov = "cluster", data = dengue_df_biannual_buffered)
#summary(dengue_biannual_model)

df <- as.data.frame(dengue_biannual_model$coeftable)
df <- df[1:45,]
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$biannual_date <- as.Date(c(unique(dengue_df_biannual$biannual_date)[c(1:16,18:46)]))
df$rainy <- 0
df$rainy[c(seq(2, 16, 2), seq(17, 45, 2))] <- 1
df$rainy <- as.character(df$rainy)
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
sfig6a <- ggplot(df) +
  geom_hline(aes(yintercept=0), colour='red', size=.4) +
  geom_errorbar(aes(x=biannual_date, ymax=upper, ymin=lower, colour=rainy), width=0, size=0.5) +
  geom_vline(aes(xintercept=as.Date("2008-04-01")), linetype='dashed', size=0.4) +
  geom_point(aes(x=as.Date("2008-04-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(biannual_date, estimate, fill=rainy), size=3, shape=21) +
  scale_fill_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  scale_colour_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  xlab("Biannual time period") + ylab("Percent\nchange\ndengue\nincidence") + 
  theme_bw()+
  scale_y_continuous(labels = scales::percent, limits=c(-2,25)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=18),
        axis.title.y=element_text(size=14,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=12),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "bottom",
        strip.text.x = element_text(size = 12))
sfig6a
sfig6_legend <- get_legend(sfig6a)
sfig6a <- sfig6a + theme(legend.position = "none")

### biannual model
leish_biannual_model <- fepois(incidence ~ i(biannual_date, fivekm, ref = '2008-04-01') + sum_precip + mean_temp + log(urban) + log(ag) | cluster + biannual_date, vcov = "cluster", data = leish_df_biannual_buffered)
#summary(leish_biannual_model)

df <- as.data.frame(leish_biannual_model$coeftable)
df <- df[1:45,]
colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value')
df$biannual_date <- as.Date(c(unique(leish_df_biannual$biannual_date)[c(1:16,18:46)]))
df$rainy <- 0
df$rainy[c(seq(2, 16, 2), seq(17, 45, 2))] <- 1
df$rainy <- as.character(df$rainy)
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
sfig6b <- ggplot(df) +
  geom_hline(aes(yintercept=0), colour='red', size=.4) +
  geom_errorbar(aes(x=biannual_date, ymax=upper, ymin=lower, colour=rainy), width=0, size=0.5) +
  geom_vline(aes(xintercept=as.Date("2008-04-01")), linetype='dashed', size=0.4) +
  geom_point(aes(x=as.Date("2008-04-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(biannual_date, estimate, fill=rainy), size=3, shape=21) +
  scale_fill_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  scale_colour_manual(name="Season", values=c('#DC267F', '#648FFF'), labels=c('Dry', 'Rainy')) + 
  xlab("Biannual time period") + ylab("Percent\nchange\nleish\nincidence") + 
  theme_bw()+
  scale_y_continuous(labels = scales::percent, limits=c(-2,25)) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=18),
        axis.title.y=element_text(size=14,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=12),
        axis.title.x=element_text(size=14),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=14),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 12))
sfig6b

#####################
## SFig 6all
#####################

sfig6all <- grid.arrange(sfig6a, sfig6b, sfig6_legend,
                         ncol = 1, nrow = 3,
                         layout_matrix = rbind(c(1),c(2),c(3)), 
                         heights=c(5,5,1))

sfig6all <- as_ggplot(sfig6all) +                                
  draw_plot_label(label = c("A", "B"), size = 14,
                  x = c(0.194, 0.194), y = c(0.99, 0.535)) 
sfig6all

ggsave("SFig6.pdf", plot=sfig6all, path="~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 8, height = 11, units="in", device = "pdf")
```

```{r, SFig S7}
#####################
## SFig 7a
#####################
#map of all 100 healthcare centers
center_lat_long <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/DIRESA_E_Salud_Coordinates_Key - best_choice.csv")
center_lat_long <- data.frame(center_lat_long)
center_lat_long <- center_lat_long %>%
  mutate(latitude= as.numeric(latitude),longitude= as.numeric(longitude)) %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))

districts <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/Madre_de_Dios.shp")
districts <- st_as_sf(districts) 
districts$geometry <- st_transform(districts$geometry, 4326)
mdd_region <- st_union(districts$geometry)

rivers <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_rivers2.shp")
rivers <- st_as_sf(rivers, crs=4326) 
rivers$geometry <- st_transform(rivers$geometry, "EPSG:4326")

roads <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_roads.shp")
roads <- st_as_sf(roads) 
roads$geometry <- st_transform(roads$geometry, 4326)
roads_mdd <- st_covers(mdd_region,roads$geometry, sparse = FALSE)
roads_mdd <- roads[roads_mdd[1,],]

highway <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/peru_roads_important.shp")
highway <- highway[which(highway$ref=="PE-30C"),]
highway <- st_as_sf(highway) 
highway$geometry <- st_transform(highway$geometry, 4326)

highway_mdd <- st_covers(mdd_region,highway$geometry, sparse = FALSE)
highway_mdd <- highway[highway_mdd[1,],]

no_axis <- theme(axis.title=element_blank(),
                 axis.text=element_blank(),
                 axis.ticks=element_blank(),
                 panel.grid.major = element_blank())

sfig7a <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#a6a6a6', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='lightblue'), linewidth=0.2, show.legend = "line") +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='#a6a6a6'), linewidth=0.5, show.legend = "line") +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='black'), linewidth=0.7, show.legend = "line") +
  geom_sf(data = center_lat_long, aes(geometry = geometry), fill='#648FFF', color='black', pch=21, size = 2, alpha=0.7) + #colour='#25CED1',
  scale_color_manual(name="", values=c('#a6a6a6','black','lightblue'),
                     labels=c('Unpaved Roads','Highway','Rivers')) +
  theme_minimal() +
  no_axis +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='bottom') +
  guides(fill=guide_legend(override.aes=list(shape=21))) +
  annotation_scale(location = "bl",height = unit(0.1, "cm"))

sfig7a
sfig7_legend <- get_legend(sfig7a)
sfig7a <- sfig7a + theme(legend.position = "none")

#####################
## SFig 7b
#####################
#map of 1 healthcare center clusters w spatial boundaries
cluster_lat_long <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_7.5km/cluster_centroids_7.5km.csv")
cluster_lat_long <- data.frame(cluster_lat_long)
cluster_lat_long <- cluster_lat_long %>%
  mutate(latitude= as.numeric(latitude),longitude= as.numeric(longitude)) %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))
#load spatial boundaries
cachement_areas <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_7.5km/diresa_nn_raster_7.5kmclus_7.5kmbuff_20kgam_00sens.shp")
cachement_areas <- st_as_sf(cachement_areas)

sfig7b <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#a6a6a6', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='lightblue'), linewidth=0.2, show.legend = "line") +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='#a6a6a6'), linewidth=0.5, show.legend = "line") +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='black'), linewidth=0.7, show.legend = "line") +
  geom_sf(data = cachement_areas, aes(geometry = geometry), fill = 'lightgrey', color='black', alpha=0.5) +
  geom_sf(data = cluster_lat_long, aes(geometry = geometry), fill='#E04490', color='black', pch=21, size = 2, alpha=0.7) + #colour='#25CED1',
  scale_color_manual(name="", values=c('#a6a6a6','black','lightblue'),
                     labels=c('Unpaved Roads','Highway','Rivers')) +
  theme_minimal() +
  no_axis +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='none') +
  guides(fill=guide_legend(override.aes=list(shape=21))) +
  annotation_north_arrow(location = "br",style = north_arrow_orienteering(text_size = 6), 
                         height = unit(0.7, "cm"), width = unit(0.7, "cm"))
sfig7b

sfig7all <- grid.arrange(sfig7a, sfig7b, sfig7_legend,                     
                         ncol = 6, nrow = 2,
                         layout_matrix = rbind(c(1,2),c(1,2),c(1,2),c(1,2),c(1,2),c(3,3)))

sfig7all <- as_ggplot(sfig7all) +                                
  draw_plot_label(label = c("A", "B"), size = 17,
                  x = c(0.01, 0.5), y = c(0.99, 0.99)) 

sfig7all


ggsave("SFig7.pdf", plot=sfig7all, "~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 9, height = 5, units = c("in"), device="pdf")
```

```{r, SFig S8}
#####################
## Load alternate clustering data and process
#####################
### 6km // 78 units
dengue_df_yearly_6km <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_6km/dengue_yearly_full_dataset_c.csv")
dengue_df_yearly_6km$year <- as.factor(dengue_df_yearly_6km$year)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_yearly_6km$incidence <- (dengue_df_yearly_6km$yearly_cases+1)/dengue_df_yearly_6km$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_yearly_6km$urban <- dengue_df_yearly_6km$urban + 0.001
dengue_df_yearly_6km$ag <- dengue_df_yearly_6km$ag + 0.001

# double check complete cases
dengue_df_yearly_6km <- dengue_df_yearly_6km[complete.cases(dengue_df_yearly_6km),]

# remove extreme outliers (all most likely due to inaccurate pop data)
dengue_df_yearly_6km <- dengue_df_yearly_6km[which(dengue_df_yearly_6km$incidence < 0.5),]

# remove PM
dengue_df_yearly_6km_no_pm <- dengue_df_yearly_6km[!(dengue_df_yearly_6km$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_yearly_6km_buffered <- dengue_df_yearly_6km[which(dengue_df_yearly_6km$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
dengue_df_yearly_6km_buffered_no_pm <- dengue_df_yearly_6km_buffered[!(dengue_df_yearly_6km_buffered$cluster %in% 1),]

### 5km // 81 units
dengue_df_yearly_5km <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_5km/dengue_yearly_full_dataset_c.csv")
dengue_df_yearly_5km$year <- as.factor(dengue_df_yearly_5km$year)

# cases + 1 so 0's dont go to infinity when logged
dengue_df_yearly_5km$incidence <- (dengue_df_yearly_5km$yearly_cases+1)/dengue_df_yearly_5km$population 

# land use vars + 0.001 so 0's dont go to infinity when logged
dengue_df_yearly_5km$urban <- dengue_df_yearly_5km$urban + 0.001
dengue_df_yearly_5km$ag <- dengue_df_yearly_5km$ag + 0.001

# double check complete cases
dengue_df_yearly_5km <- dengue_df_yearly_5km[complete.cases(dengue_df_yearly_5km),]

# remove extreme outliers (all most likely due to inaccurate pop data)
dengue_df_yearly_5km <- dengue_df_yearly_5km[which(dengue_df_yearly_5km$incidence < 0.5),]

# remove PM
dengue_df_yearly_5km_no_pm <- dengue_df_yearly_5km[!(dengue_df_yearly_5km$cluster %in% 1),]

# build final, buffered dataset where below 5km is 1, between 5km and 10km are thrown out (buffer zone), and above 10km is 0
dengue_df_yearly_5km_buffered <- dengue_df_yearly_5km[which(dengue_df_yearly_5km$all_cutoffs %in% c(1,2,4,5,6,7,0)),]
dengue_df_yearly_5km_buffered_no_pm <- dengue_df_yearly_5km_buffered[!(dengue_df_yearly_5km_buffered$cluster %in% 1),]

#####################
## SFig 8a
#####################

## main specification
dengue_yearly_model_main <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_buffered)

dengue_yearly_model_6km <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_6km_buffered)

dengue_yearly_model_5km <- fepois(incidence ~ i(year, fivekm, ref = '2008-01-01') + log(urban) + log(ag) + sum_precip + mean_temp | cluster + year, vcov = "cluster", data = dengue_df_yearly_5km_buffered)

df <- rbind(
  as.data.frame(cbind(dengue_yearly_model_5km$coeftable[1:22,],c(rep("5km clustering",22)))),
  as.data.frame(cbind(dengue_yearly_model_6km$coeftable[1:22,],c(rep("6km clustering",22)))),
  as.data.frame(cbind(dengue_yearly_model_main$coeftable[1:22,],c(rep("7.5km clustering",22)))))

colnames(df) <- c('estimate', 'std_error', 't_value', 'p_value', 'Model')
df$year <- c(seq(as.Date("2000-01-01"), as.Date("2007-01-01"), by="year"),
             seq(as.Date("2009-01-01"), as.Date("2022-01-01"), by="year"))
df$estimate <- as.numeric(df$estimate)
df$std_error <- as.numeric(df$std_error)
df$upper <- df$estimate+1.96*df$std_error
df$lower <- df$estimate-1.96*df$std_error
df$estimate <- exp(df$estimate)-1
df$upper <- exp(df$upper)-1
df$lower <- exp(df$lower)-1
df$Model <- factor(as.character(df$Model), levels=c('5km clustering','6km clustering','7.5km clustering'))

sfig8a <- ggplot(df, aes(group=Model)) +
  geom_hline(aes(yintercept=0), colour='red', linewidth=.4) +
  geom_errorbar(aes(x=year, ymax=upper, ymin=lower, color=Model), width=0.1, linewidth=0.5, alpha=0.8) +
  geom_vline(aes(xintercept=as.Date("2008-01-01")), linetype='dashed', linewidth=0.4) +
  geom_point(aes(x=as.Date("2008-01-01"), y=0), size=3, shape=21, fill='white') +
  geom_point(aes(year, estimate, fill=Model), color='black', size=3, shape=21, alpha=0.7) +
  scale_color_manual(name = "Clustering specification", values=list('5km clustering'="#648FFF", '6km clustering'='#FFD218','7.5km clustering'='#DC267F'),
                     labels=list('5km clustering (n = 84)','6km clustering (n = 78)','7.5km clustering (n = 71)'))+
  scale_fill_manual(name = "Clustering specification", values=list('5km clustering'="#648FFF", '6km clustering'='#FFD218','7.5km clustering'='#DC267F'),
                     labels=list('5km clustering (n = 84)','6km clustering (n = 78)','7.5km clustering (n = 71)'))+
  xlab("Year") + ylab("Percent\nchange\ndengue\nincidence") + 
  theme_bw()+
  scale_y_continuous(labels = scales::percent) +
  theme(plot.title = element_text(size=20),
        plot.title.position = "plot",
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title=element_text(size=18),
        axis.title.y=element_text(size=12,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=12),
        axis.text.x=element_text(size=12),
        axis.text = element_text(size=12),
        legend.text=element_text(size=10),
        legend.title=element_text(size=10),
        legend.position = "right",
        strip.text.x = element_text(size = 12))
sfig8a

#####################
## SFig 8b
#####################
cluster_lat_long_7.5km <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_7.5km/cluster_centroids_7.5km.csv")
cluster_lat_long_7.5km <- data.frame(cluster_lat_long_7.5km)
cluster_lat_long_7.5km <- cluster_lat_long_7.5km %>%
  mutate(latitude= as.numeric(latitude),longitude= as.numeric(longitude)) %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))

districts <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/Madre_de_Dios.shp")
districts <- st_as_sf(districts) 
districts$geometry <- st_transform(districts$geometry, 4326)
mdd_region <- st_union(districts$geometry)

rivers <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_rivers2.shp")
rivers <- st_as_sf(rivers, crs=4326) 
rivers$geometry <- st_transform(rivers$geometry, "EPSG:4326")

roads <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/mdd_roads.shp")
roads <- st_as_sf(roads) 
roads$geometry <- st_transform(roads$geometry, 4326)
roads_mdd <- st_covers(mdd_region,roads$geometry, sparse = FALSE)
roads_mdd <- roads[roads_mdd[1,],]

highway <- read_sf("~/Desktop/doctorate/ch2 mdd highway/data/shapefiles/peru_roads_important.shp")
highway <- highway[which(highway$ref=="PE-30C"),]
highway <- st_as_sf(highway) 
highway$geometry <- st_transform(highway$geometry, 4326)

highway_mdd <- st_covers(mdd_region,highway$geometry, sparse = FALSE)
highway_mdd <- highway[highway_mdd[1,],]

no_axis <- theme(axis.title=element_blank(),
                 axis.text=element_blank(),
                 axis.ticks=element_blank(),
                 panel.grid.major = element_blank())

sfig8b <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#a6a6a6', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='lightblue'), linewidth=0.2, show.legend = "line") +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='#a6a6a6'), linewidth=0.5, show.legend = "line") +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='black'), linewidth=0.7, show.legend = "line") +
  geom_sf(data = cluster_lat_long_7.5km, aes(geometry = geometry), fill='#DC267F', color='black', pch=21, size = 2, alpha=0.7) +
  scale_color_manual(name="", values=c('#a6a6a6','black','lightblue'),
                     labels=c('Unpaved Roads','Highway','Rivers')) +
  theme_minimal() +
  no_axis +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='none') +
  guides(fill=guide_legend(override.aes=list(shape=21))) +
  annotation_scale(location = "bl",height = unit(0.08, "cm"))

sfig8b
#sfig8_legend <- get_legend(sfig8b)
#sfig8b <- sfig8b + theme(legend.position = "none")

#####################
## SFig 8c
#####################
cluster_lat_long_6km <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_6km/cluster_centroids_6km.csv")
cluster_lat_long_6km <- data.frame(cluster_lat_long_6km)
cluster_lat_long_6km <- cluster_lat_long_6km %>%
  mutate(latitude= as.numeric(latitude),longitude= as.numeric(longitude)) %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))

sfig8c <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#a6a6a6', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='lightblue'), linewidth=0.2, show.legend = "line") +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='#a6a6a6'), linewidth=0.5, show.legend = "line") +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='black'), linewidth=0.7, show.legend = "line") +
  geom_sf(data = cluster_lat_long_6km, aes(geometry = geometry), fill='#FFD218', color='black', pch=21, size = 2, alpha=0.7) +
  scale_color_manual(name="", values=c('#a6a6a6','black','lightblue'),
                     labels=c('Unpaved Roads','Highway','Rivers')) +
  theme_minimal() +
  no_axis +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='none') +
  guides(fill=guide_legend(override.aes=list(shape=21)))

sfig8c

#####################
## SFig 8d
#####################
cluster_lat_long_5km <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/spatial_units/buffer_5km/cluster_centroids_5km.csv")
cluster_lat_long_5km <- data.frame(cluster_lat_long_5km)
cluster_lat_long_5km <- cluster_lat_long_5km %>%
  mutate(latitude= as.numeric(latitude),longitude= as.numeric(longitude)) %>%
  sf::st_as_sf(coords = c('longitude','latitude'), crs = st_crs(4326))

sfig8d <- ggplot() +
  geom_sf(data = mdd_region, fill='#ffffff', color='#a6a6a6', size=.15, show.legend = FALSE) +
  geom_sf(data = rivers, aes(geometry = geometry, color='lightblue'), linewidth=0.2, show.legend = "line") +
  geom_sf(data = roads_mdd, aes(geometry = geometry, color='#a6a6a6'), linewidth=0.5, show.legend = "line") +
  geom_sf(data = highway_mdd, aes(geometry = geometry, color='black'), linewidth=0.7, show.legend = "line") +
  geom_sf(data = cluster_lat_long_5km, aes(geometry = geometry), fill='#648FFF', color='black', pch=21, size = 2, alpha=0.7) +
  scale_color_manual(name="", values=c('#a6a6a6','black','lightblue'),
                     labels=c('Unpaved Roads','Highway','Rivers')) +
  theme_minimal() +
  no_axis +
  theme(legend.text=element_text(size=12),
        legend.title=element_text(size=14),
        legend.position='none') +
  guides(fill=guide_legend(override.aes=list(shape=21))) +
  annotation_north_arrow(location = "br",style = north_arrow_orienteering(text_size = 6), 
                         height = unit(0.5, "cm"), width = unit(0.5, "cm"))

sfig8d

#####################
## SFig 8all
#####################
sfig8all <- grid.arrange(sfig8b, sfig8c, sfig8d, sfig8a,                     
                         ncol = 4, nrow = 2,
                         layout_matrix = rbind(c(1,2,3),c(1,2,3),c(1,2,3),c(1,2,3),c(5,5,5),c(5,5,5),c(5,5,5),c(5,5,5),c(5,5,5)))

sfig8all <- as_ggplot(sfig8all) +                                
  draw_plot_label(label = c("A", "B", "C", "D"), size = 17,
                  x = c(0.01, 0.33, 0.66, 0.01), y = c(0.95, 0.95, 0.95, 0.57)) 

sfig8all

ggsave("SFig8.pdf", plot=sfig8all, "~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 9, height = 8, units = c("in"), device="pdf")
```

```{r, SFig S9}
#####################
## SFig 9a
#####################
### precip data
precip_monthly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/covariates_data_7.5km/mdd_precipitation_monthly_sum.csv")
precip_monthly <- precip_monthly[,c(2:278)]
precip_monthly <- precip_monthly[,c(277,1:276)]
colnames(precip_monthly)[2:277] <- as.character(seq(as.Date("2000-01-01"), as.Date("2022-12-01"), by="months"))
colnames(precip_monthly)[1] <- "cluster"
precip_monthly_mdd_long <- precip_monthly %>%
  pivot_longer(cols = c(2:277), 
               names_to = "month", 
               values_to = "sum_precip")
precip_monthly_mdd_long$year <- format(as.Date(precip_monthly_mdd_long$month, format="%Y-%m-%d"),"%Y")

#look at precip to inform rainy season/biannual split
precip_monthly_mdd_long$year <- format(as.Date(precip_monthly_mdd_long$month, format="%Y-%m-%d"),"%Y")
precip_monthly_all <- precip_monthly_mdd_long %>%
  group_by(month, year) %>%
  summarize(sum_precip = sum(sum_precip)) 
precip_monthly_all$month_wo_year <- format(as.Date(precip_monthly_all$month, format="%Y-%m-%d"),"%m")
sfig9a <- ggplot(precip_monthly_all) +
  geom_line(aes(x=month_wo_year,y=sum_precip,group=year), color="black", alpha=0.7) +
  geom_hline(yintercept=mean(precip_monthly_all$sum_precip), color='red', linetype="dashed") +
  geom_vline(xintercept=04, color='red', linetype="dashed") +
  geom_vline(xintercept=10, color='red', linetype="dashed") +
  xlab("Month") + ylab("Acc.\nmonthly\nprecip") + 
  theme_bw()+
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=10,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=10),
        axis.text.x=element_text(size=10, angle=45, vjust=0.45),
        axis.text = element_text(size=16),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 14))
sfig9a

#####################
## SFig 9b
#####################
dengue_monthly <- read.csv("~/Desktop/doctorate/ch2 mdd highway/data/processed_case_data_7.5km/dengue_monthly_full_dataset.csv")
dengue_monthly <- dengue_monthly %>%
  group_by(month, year) %>%
  summarize(mean_cases = mean(monthly_cases))
dengue_monthly$month_wo_year <- format(as.Date(dengue_monthly$month, format="%Y-%m-%d"),"%m")
dengue_monthly$year_wo_month <- format(as.Date(dengue_monthly$month, format="%Y-%m-%d"),"%Y")

sfig9b <- ggplot(dengue_monthly) +
  geom_line(aes(x=month_wo_year,y=mean_cases,group=year), color='black', alpha=0.7) +
  #geom_hline(yintercept=mean(dengue_monthly$mean_cases), color='red', linetype="dashed") +
  geom_vline(xintercept=04, color='red', linetype="dashed") +
  geom_vline(xintercept=10, color='red', linetype="dashed") +
  xlab("Month") + ylab("Mean\nmonthly\ndengue\ncases") + 
  theme_bw()+
  theme(plot.title = element_text(hjust=0.5, size=26, face="italic"),
        plot.subtitle = element_text(hjust=0.5, size=22),
        axis.title.y=element_text(size=10,angle=0, vjust=.5, hjust=0.5),
        axis.text.y=element_text(size=10),
        axis.title.x=element_text(size=10),
        axis.text.x=element_text(size=10, angle=45, vjust=0.45),
        axis.text = element_text(size=16),
        legend.text=element_text(size=12),
        legend.title=element_text(size=12),
        legend.position = "none",
        strip.text.x = element_text(size = 14)) +
  geom_text_repel(
    data = subset(dengue_monthly, (month_wo_year == 12 & year_wo_month %in% c('2019','2022','2010','2020','2012','2009','2013','2015'))),
    aes(x = month_wo_year, y = mean_cases, label = year_wo_month), 
    color='slategrey',
    point.size = NA,
    min.segment.length = 0,
    size = 2.5,
    hjust = 'right',
    direction = "y",
    nudge_x = 0.8,
    segment.color = 'slategrey',
    segment.alpha = 0.7,
    #segment.curvature = -0.1,
    #segment.ncp = 3,
    #segment.angle = 20,
    point.padding = 0.1,
    show.legend = FALSE,
    angle = 0,
    arrow = arrow(length = unit(0.01, "npc")),
    expand = expansion(mult = 0.4)
  )
sfig9b

#####################
## SFig 9all
#####################
sfig9all <- grid.arrange(sfig9a, sfig9b,
                         ncol = 1, nrow = 2,
                         layout_matrix = rbind(c(1),c(2)), 
                         heights=c(1,1))

sfig9all <- as_ggplot(sfig9all) +                                
  draw_plot_label(label = c("A", "B"), size = 14,
                  x = c(0.04, 0.04), y = c(1, 0.50)) 
sfig9all

ggsave("SFig9.pdf", plot=sfig9all, path="~/Desktop/doctorate/ch2 mdd highway/supplementary_figures/", width = 8, height = 8, units="in", device = "pdf")
```
